Built-in material allowing the user to test the package with demo material

Built-in tibbles and lists allowing the user to test the package with demo material.

DEMO_files

Format

list

A list with 22 elements used for testing purpose (data-frames and lists):

DEMO_data_processing_elements - final

Data processing element without error

DEMO_data_processing_elements - with error

Data processing element containing errors

DEMO_data_processing_elements - work in progress

Data processing element in construction

DEMO_dataschema

DataSchema used, combined with data processing elements and study-specific material

dd_MELBOURNE_1_format_maelstrom

Data dictionary (1) of Melbourne dataset

dd_MELBOURNE_2_format_maelstrom

Data dictionary (2) of Melbourne dataset

dd_PARIS_format_maelstrom

Data dictionary of Paris dataset

dd_PARIS_format_preprocessed - ERROR

Data dictionary of Paris dataset containing errors

dd_PARIS_format_preprocessed

Data dictionary of Paris in preprocessed format

dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_maelstrom_tagged - ERROR

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_maelstrom_tagged

Data dictionary of Tokyo dataset

dd_TOKYO_format_opal_tagged - ERROR WITH TAXO

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_opal_tagged

Data dictionary of Tokyo dataset opal format

dataset_MELBOURNE_1

Dataset of Melbourne (1)

dataset_MELBOURNE_2

Dataset of Melbourne (2)

dataset_PARIS

Dataset of Paris

dataset_TOKYO - ERROR WITH DATA

Dataset of Tokyo containing errors

dataset_TOKYO

Dataset of Tokyo

PARIS_taxonomy

Taxonomy specific to Paris dataset

taxonomy_mlstr

Maelstrom Taxonomy

taxonomy_opal

Opal Taxonomy

...

Examples

{

DEMO_files

}
#> $`data_processing_elements - final`
#> $`data_processing_elements - final`$dataProcessing
#> # A tibble: 52 × 19
#>    index ss_dataschema_table            dataschema_variable ss_acronym ss_table 
#>    <dbl> <chr>                          <chr>               <chr>      <chr>    
#>  1     1 harmonized.dataset_MELBOURNE_1 adm_unique_id       MELBOURNE  dataset_…
#>  2     1 harmonized.dataset_MELBOURNE_2 adm_unique_id       MELBOURNE  dataset_…
#>  3     1 harmonized.dataset_PARIS       adm_unique_id       PARIS      dataset_…
#>  4     1 harmonized.dataset_TOKYO       adm_unique_id       TOKYO      dataset_…
#>  5     2 harmonized.dataset_MELBOURNE_1 adm_study           MELBOURNE  dataset_…
#>  6     2 harmonized.dataset_MELBOURNE_2 adm_study           MELBOURNE  dataset_…
#>  7     2 harmonized.dataset_PARIS       adm_study           PARIS      dataset_…
#>  8     2 harmonized.dataset_TOKYO       adm_study           TOKYO      dataset_…
#>  9     3 harmonized.dataset_MELBOURNE_1 adm_year_dce        MELBOURNE  dataset_…
#> 10     3 harmonized.dataset_MELBOURNE_2 adm_year_dce        MELBOURNE  dataset_…
#> # ℹ 42 more rows
#> # ℹ 14 more variables: ss_variables <chr>, ss_format <lgl>,
#> #   ss_time_period <lgl>, valueType <chr>, ss_time_collection <lgl>,
#> #   ss_source_information <lgl>, ss_informant <lgl>, ss_keywords <lgl>,
#> #   comment <lgl>, harmo_status <lgl>, harmo_status_detail <lgl>,
#> #   rule_category <chr>, harmo_rule <chr>, comment_internal <lgl>
#> 
#> 
#> $`data_processing_elements - with error`
#> $`data_processing_elements - with error`$dataProcessing
#> # A tibble: 52 × 19
#>    index ss_dataschema_table            dataschema_variable ss_acronym ss_table 
#>    <dbl> <chr>                          <chr>               <chr>      <chr>    
#>  1     1 harmonized.dataset_MELBOURNE_1 adm_unique_id       MELBOURNE  dataset_…
#>  2     1 harmonized.dataset_MELBOURNE_2 adm_unique_id       MELBOURNE  dataset_…
#>  3     1 harmonized.dataset_PARIS       adm_unique_id       PARIS      dataset_…
#>  4     1 harmonized.dataset_TOKYO       adm_unique_id       TOKYO      dataset_…
#>  5     2 harmonized.dataset_MELBOURNE_1 adm_study           MELBOURNE  dataset_…
#>  6     2 harmonized.dataset_MELBOURNE_2 adm_study           MELBOURNE  dataset_…
#>  7     2 harmonized.dataset_PARIS       adm_study           PARIS      dataset_…
#>  8     2 harmonized.dataset_TOKYO       adm_study           TOKYO      dataset_…
#>  9     3 harmonized.dataset_MELBOURNE_1 adm_year_dce        MELBOURNE  dataset_…
#> 10     3 harmonized.dataset_MELBOURNE_2 adm_year_dce        MELBOURNE  dataset_…
#> # ℹ 42 more rows
#> # ℹ 14 more variables: ss_variables <chr>, ss_format <lgl>,
#> #   ss_time_period <lgl>, valueType <chr>, ss_time_collection <lgl>,
#> #   ss_source_information <lgl>, ss_informant <lgl>, ss_keywords <lgl>,
#> #   comment <lgl>, harmo_status <lgl>, harmo_status_detail <lgl>,
#> #   rule_category <chr>, harmo_rule <chr>, comment_internal <lgl>
#> 
#> 
#> $`data_processing_elements - work in progress`
#> $`data_processing_elements - work in progress`$dataProcessing
#> # A tibble: 52 × 19
#>    index ss_dataschema_table            dataschema_variable ss_acronym ss_table 
#>    <dbl> <chr>                          <chr>               <chr>      <chr>    
#>  1     1 harmonized.dataset_MELBOURNE_1 adm_unique_id       MELBOURNE  dataset_…
#>  2     2 harmonized.dataset_MELBOURNE_2 adm_unique_id       MELBOURNE  dataset_…
#>  3     3 harmonized.dataset_PARIS       adm_unique_id       PARIS      dataset_…
#>  4     4 harmonized.dataset_TOKYO       adm_unique_id       TOKYO      dataset_…
#>  5     5 harmonized.dataset_MELBOURNE_1 adm_study           MELBOURNE  dataset_…
#>  6     6 harmonized.dataset_MELBOURNE_2 adm_study           MELBOURNE  dataset_…
#>  7     7 harmonized.dataset_PARIS       adm_study           PARIS      dataset_…
#>  8     8 harmonized.dataset_TOKYO       adm_study           TOKYO      dataset_…
#>  9     9 harmonized.dataset_MELBOURNE_1 adm_year_dce        MELBOURNE  dataset_…
#> 10    10 harmonized.dataset_MELBOURNE_2 adm_year_dce        MELBOURNE  dataset_…
#> # ℹ 42 more rows
#> # ℹ 14 more variables: ss_variables <chr>, ss_format <lgl>,
#> #   ss_time_period <lgl>, valueType <chr>, ss_time_collection <lgl>,
#> #   ss_source_information <lgl>, ss_informant <lgl>, ss_keywords <lgl>,
#> #   comment <lgl>, harmo_status <lgl>, harmo_status_detail <lgl>,
#> #   rule_category <chr>, harmo_rule <chr>, comment_internal <lgl>
#> 
#> 
#> $`dataschema - final`
#> $`dataschema - final`$Variables
#> # A tibble: 13 × 27
#>    index project table             name  `label:en` valueType unit  other_format
#>    <dbl> <chr>   <chr>             <chr> <chr>      <chr>     <chr> <lgl>       
#>  1     1 DEMO    harmonized_study… adm_… Unique id… text      NA    NA          
#>  2     2 DEMO    harmonized_study… adm_… Indicator… text      NA    NA          
#>  3     3 DEMO    harmonized_study… adm_… Indicator… text      NA    NA          
#>  4     4 DEMO    harmonized_study… sdc_… Participa… integer   years NA          
#>  5     5 DEMO    harmonized_study… sdc_… Gender of… integer   NA    NA          
#>  6     6 DEMO    harmonized_study… phy_… participa… decimal   cm    NA          
#>  7     7 DEMO    harmonized_study… phy_… participa… decimal   kg    NA          
#>  8     8 DEMO    harmonized_study… phy_… participa… decimal   kg/m  NA          
#>  9     9 DEMO    harmonized_study… rep_… whether t… integer   NA    NA          
#> 10    10 DEMO    harmonized_study… rep_… whether t… integer   NA    NA          
#> 11    11 DEMO    harmonized_study… lsb_… whether t… integer   NA    NA          
#> 12    12 DEMO    harmonized_study… lsb_… whether t… integer   NA    NA          
#> 13    13 DEMO    harmonized_study… lsb_… participa… integer   NA    NA          
#> # ℹ 19 more variables: individual_entity <lgl>, time_period <lgl>,
#> #   source <lgl>, informant <lgl>, time_collection <lgl>, wording <lgl>,
#> #   measures <lgl>, procedures <lgl>, instructions <lgl>, comments <lgl>,
#> #   `Mlstr_additional::Source` <lgl>, `Mlstr_additional::Target` <lgl>,
#> #   `Mlstr_area::1` <chr>, `Mlstr_area::1.term` <chr>,
#> #   `Mlstr_area::1.scale` <chr>, `Mlstr_area::2` <lgl>,
#> #   `Mlstr_area::2.term` <lgl>, `Mlstr_area::3` <lgl>, …
#> 
#> $`dataschema - final`$Categories
#> # A tibble: 13 × 6
#>    project table                            variable    `label:en`  name missing
#>    <chr>   <chr>                            <chr>       <chr>      <dbl>   <dbl>
#>  1 DEMO    harmonized_study_DEMO_table_DEMO sdc_gender  Male           1       0
#>  2 DEMO    harmonized_study_DEMO_table_DEMO sdc_gender  Female         2       0
#>  3 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_e… never pre…     0       0
#>  4 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_e… pregnant …     1       0
#>  5 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_c… currently…     0       0
#>  6 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_c… not curre…     1       0
#>  7 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_ev… never smo…     0       0
#>  8 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_ev… smoked on…     1       0
#>  9 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_cu… currently…     0       0
#> 10 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_cu… not curre…     1       0
#> 11 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… never smo…     0       0
#> 12 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… former sm…     1       0
#> 13 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… current s…     2       0
#> 
#> 
#> $dd_MELBOURNE_1_format_maelstrom
#> $dd_MELBOURNE_1_format_maelstrom$Variables
#> # A tibble: 4 × 7
#>   index name   `label:en` `description:en`      valueType unit   other_format
#>   <dbl> <chr>  <chr>      <chr>                 <chr>     <chr>  <lgl>       
#> 1     1 id     id         id of the participant text      NA     NA          
#> 2     2 Gender Gender     Gender                integer   NA     NA          
#> 3     3 BMI    BMI        Body Mass Index       decimal   kg/m^2 NA          
#> 4     4 age    age        Age of Participant    integer   years  NA          
#> 
#> $dd_MELBOURNE_1_format_maelstrom$Categories
#> # A tibble: 3 × 4
#>   variable  name `label:en`           missing
#>   <chr>    <dbl> <chr>                <lgl>  
#> 1 age       -888 don't want to answer TRUE   
#> 2 Gender       1 Male                 FALSE  
#> 3 Gender       2 Female               FALSE  
#> 
#> 
#> $dd_MELBOURNE_2_format_maelstrom
#> $dd_MELBOURNE_2_format_maelstrom$Variables
#> # A tibble: 3 × 7
#>   index name       `label:en` `description:en`      valueType unit  other_format
#>   <dbl> <chr>      <chr>      <chr>                 <chr>     <lgl> <lgl>       
#> 1     1 id         id         id of the participant text      NA    NA          
#> 2     2 smo_status smo_curr   Whether the particip… integer   NA    NA          
#> 3     3 prg_curr   prg_curr   Are you currently pr… integer   NA    NA          
#> 
#> $dd_MELBOURNE_2_format_maelstrom$Categories
#> # A tibble: 9 × 4
#>   variable    name `label:en`             missing
#>   <chr>      <dbl> <chr>                  <lgl>  
#> 1 prg_curr       0 not currently pregnant FALSE  
#> 2 prg_curr       1 currently pregnant     FALSE  
#> 3 prg_curr       8 Don’t want to answer   TRUE   
#> 4 prg_curr       9 Don’t know             TRUE   
#> 5 prg_curr     -77 not applicable         TRUE   
#> 6 smo_status     1 never smoked           FALSE  
#> 7 smo_status     2 current smoker         FALSE  
#> 8 smo_status     3 former smoker          FALSE  
#> 9 smo_status   -77 don't want to answer   TRUE   
#> 
#> 
#> $`dd_PARIS_format_flatten - ERROR`
#> $`dd_PARIS_format_flatten - ERROR`$Variables
#> # A tibble: 7 × 11
#>   index name     `label:fr`      valueType unit    other_format `description:fr`
#>   <dbl> <chr>    <chr>           <chr>     <chr>   <lgl>        <chr>           
#> 1     1 ID       id              text      NA      NA           id du participa…
#> 2     2 SEX      Sexe            integer   NA      NA           Sexe du partici…
#> 3     3 BMI      IMC             decimal   kg/m^2  NA           Indice de Masse…
#> 4     4 AGE      Age             integer   years   NA           Age du Particip…
#> 5     5 SMO      Fumeur          integer   NA      NA           Fumeur regulier 
#> 6     6 SMO_QTY  fum_qte_semaine integer   cigare… NA           nombre de cigar…
#> 7     7 PRG_EVER enceinte_occur  integer   NA      NA           Si la personne …
#> # ℹ 4 more variables: `spss::description:en` <chr>, `spss::label` <chr>,
#> #   `spss::term` <chr>, `Categories::label` <chr>
#> 
#> 
#> $dd_PARIS_format_flatten
#> $dd_PARIS_format_flatten$Variables
#> # A tibble: 7 × 19
#>   name     `label:fr`    `description:fr` valueType unit  Mlstr_additional::So…¹
#>   <chr>    <chr>         <chr>            <chr>     <chr> <chr>                 
#> 1 ID       id            id du participa… text      NA    Questionnaire         
#> 2 SEX      Sexe          Sexe du partici… integer   NA    Questionnaire         
#> 3 BMI      IMC           Indice de Masse… decimal   kg/m… Physical_measure      
#> 4 AGE      Age           Age du Particip… integer   years Physical_measure      
#> 5 SMO      Fumeur        Fumeur regulier  integer   NA    Physical_measure      
#> 6 SMO_QTY  fum_qte_sema… nombre de cigar… integer   ciga… Questionnaire         
#> 7 PRG_EVER enceinte_occ… Si la personne … integer   NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 13 more variables: `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::Administrative_information` <chr>,
#> #   `Mlstr_area::Sociodemographic_economic_characteristics` <chr>,
#> #   `Mlstr_area::Physical_measures` <chr>, `Mlstr_area::Reproduction` <chr>,
#> #   `Mlstr_area::Lifestyle_behaviours` <chr>,
#> #   `Mlstr_cogscale::Cog_scale` <chr>, `Mlstr_habits::Tobacco` <chr>, …
#> 
#> 
#> $dd_PARIS_format_maelstrom
#> $dd_PARIS_format_maelstrom$Variables
#> # A tibble: 7 × 9
#>   index name     `label:fr` `description:fr` valueType unit  spss_description_en
#>   <dbl> <chr>    <chr>      <chr>            <chr>     <chr> <chr>              
#> 1     1 ID       id         id du participa… text      NA    id of the particip…
#> 2     2 SEX      Sexe       Sexe du partici… integer   NA    Gender             
#> 3     3 BMI      IMC        Indice de Masse… decimal   kg/m… Body Mass Index    
#> 4     4 AGE      Age        Age du Particip… integer   years Age of Participant 
#> 5     5 SMO      Fumeur     Fumeur regulier  integer   NA    Regular smoker     
#> 6     6 SMO_QTY  fum_qte_s… nombre de cigar… integer   ciga… number of cigarett…
#> 7     7 PRG_EVER enceinte_… Si la personne … integer   NA    whether the partic…
#> # ℹ 2 more variables: spss_label <chr>, spss_comment <chr>
#> 
#> $dd_PARIS_format_maelstrom$Categories
#> # A tibble: 9 × 4
#>   variable  name `label:fr`                 missing
#>   <chr>    <dbl> <chr>                      <lgl>  
#> 1 SEX          0 Homme                      FALSE  
#> 2 SEX          1 Femme                      FALSE  
#> 3 SMO          0 Non-Fumeur                 FALSE  
#> 4 SMO          1 Fumeur                     FALSE  
#> 5 SMO_QTY     -8 SKIP PATTERN               FALSE  
#> 6 PRG_EVER     0 Jamais enceinte            FALSE  
#> 7 PRG_EVER     1 Enceinte au moins une fois FALSE  
#> 8 PRG_EVER     9 NSP                        TRUE   
#> 9 PRG_EVER    -8 SKIP PATTERN               TRUE   
#> 
#> 
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA`
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA`$Variables
#> # A tibble: 9 × 15
#>   index name  `label:en` `description:en` valueType unit  Mlstr_additional::So…¹
#>   <dbl> <chr> <chr>      <chr>            <chr>     <chr> <chr>                 
#> 1     1 part… id of the… id of the parti… text      NA    Questionnaire         
#> 2     2 gndr  gndr       gender of the p… boolean   NA    Questionnaire         
#> 3     3 heig… height     height of the p… decimal   cm    Physical_measure      
#> 4     4 weig… weight_ms  weight of the p… integer   kg    Physical_measure      
#> 5     5 weig… weight_dc  weight of the p… integer   kg    NA                    
#> 6     6 dob   dob        date of birth o… text      years Participant           
#> 7     7 prg_… prg_ever   whether the par… integer   NA    Questionnaire         
#> 8     8 empty empty      empty column     integer   NA    Questionnaire         
#> 9     9 open… opentext   open text        boolean   NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 8 more variables: `Mlstr_additional::Target` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <chr>,
#> #   `Mlstr_area::2` <chr>, `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>
#> 
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA`$Categories
#> # A tibble: 11 × 4
#>    variable  name   `label:en`            missing
#>    <chr>     <chr>  <chr>                 <lgl>  
#>  1 gndr      Male   Male                  FALSE  
#>  2 gndr      Female Female                FALSE  
#>  3 gndr      -77    Don’t want to answer  TRUE   
#>  4 weight_ms -88    Don’t want to answer  TRUE   
#>  5 weight_ms -99    Do not remember       TRUE   
#>  6 prg_ever  0      never pregnant        FALSE  
#>  7 prg_ever  1      pregnant once or more FALSE  
#>  8 prg_ever  2      currently pregnant    FALSE  
#>  9 prg_ever  8      Don’t want to answer  TRUE   
#> 10 prg_ever  -7     not applicable        TRUE   
#> 11 dob       -88    Don’t want to answer  TRUE   
#> 
#> 
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR`
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR`$Variables
#> # A tibble: 12 × 17
#>    index table name     `label:en` `description:en` valueType `copy units` unit 
#>    <dbl> <chr> <chr>    <chr>      <chr>            <chr>     <chr>        <chr>
#>  1     1 DEMO  part_id  id of the… id of the parti… NA        NA           NA   
#>  2     2 DEMO  gndr     gndr       gender of the p… boolean   NA           NA   
#>  3     3 DEMO  height   height     height of the p… decimal   cm           cm   
#>  4     4 DEMO  weight_… weight_ms  weight of the p… integer   kg           kg   
#>  5     5 DEMO  weight … weight_dc  weight of the p… integer   kg           kg   
#>  6     5 DEMO  weight_… weight_dc  weight of the p… integr    kg           kg   
#>  7     6 DEMO  dob      NA         date of birth o… date      years        years
#>  8     7 DEMO  prg_evr  prg_ever   whether the par… integer   NA           NA   
#>  9     8 DEMO  empty    empty      empty column     integer   NA           NA   
#> 10     9 DEMO  opentext opentext   open text        character NA           NA   
#> 11    NA DEMO  opentext NA         NA               NA        NA           NA   
#> 12     8 DEMO  NA       no name    no name          integer   NA           NA   
#> # ℹ 9 more variables: `Mlstr_additional::Source` <chr>,
#> #   `Mlstr_additional::Target` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <chr>,
#> #   `Mlstr_area::2` <chr>, `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>
#> 
#> $`dd_TOKYO_format_maelstrom_tagged - ERROR`$Categories
#> # A tibble: 15 × 5
#>    table variable  name         `label:en`           missing
#>    <chr> <chr>     <chr>        <chr>                <chr>  
#>  1 DEMO  gndr      Male         Male                 FALSE  
#>  2 DEMO  gndr      Female       Female               FALSE  
#>  3 DEMO  gndr      -77          Don’t want to answer TRUE   
#>  4 DEMO  weight_ms -88          Don’t want to answer TRUE   
#>  5 DEMO  weight_ms -99          Do not remember      TRUE   
#>  6 DEMO  weight_sm -99          Don’t want to answer TRUE   
#>  7 DEMO  prg_ever  0            never pregnant       FALSE  
#>  8 DEMO  prg_ever  1            NA                   FALSE  
#>  9 DEMO  prg_ever  2            currently pregnant   FLASE  
#> 10 DEMO  NA        6            Inaccurate           TRUE   
#> 11 DEMO  prg_ever  8            Don’t want to answer TRUE   
#> 12 DEMO  prg_ever  NA           Don’t know           TRUE   
#> 13 DEMO  prg_ever  -7           not applicable       TRUE   
#> 14 DEMO  prg_ever  -7           not applicable       NA     
#> 15 DEMO  height    SKIP PATTERN SKIP PATTERN         TRUE   
#> 
#> 
#> $dd_TOKYO_format_maelstrom_tagged
#> $dd_TOKYO_format_maelstrom_tagged$Variables
#> # A tibble: 9 × 15
#>   index name  `label:en` `description:en` valueType unit  Mlstr_additional::So…¹
#>   <dbl> <chr> <chr>      <chr>            <chr>     <chr> <chr>                 
#> 1     1 part… id of the… id of the parti… text      NA    Questionnaire         
#> 2     2 gndr  gndr       gender of the p… text      NA    Questionnaire         
#> 3     3 heig… height     height of the p… integer   cm    Physical_measure      
#> 4     4 weig… weight_ms  weight of the p… integer   kg    Physical_measure      
#> 5     5 weig… weight_dc  weight of the p… decimal   kg    Physical_measure      
#> 6     6 dob   dob        date of birth o… date      years Questionnaire         
#> 7     7 prg_… prg_ever   whether the par… integer   NA    Questionnaire         
#> 8     8 empty empty      empty column     integer   NA    Questionnaire         
#> 9     9 open… opentext   open text        text      NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 8 more variables: `Mlstr_additional::Target` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <lgl>,
#> #   `Mlstr_area::2` <chr>, `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>
#> 
#> $dd_TOKYO_format_maelstrom_tagged$Categories
#> # A tibble: 11 × 4
#>    variable  name   `label:en`            missing
#>    <chr>     <chr>  <chr>                 <lgl>  
#>  1 gndr      Male   Male                  FALSE  
#>  2 gndr      Female Female                FALSE  
#>  3 gndr      -77    Don’t want to answer  TRUE   
#>  4 weight_ms -88    Don’t want to answer  TRUE   
#>  5 weight_ms -99    Don’t know            TRUE   
#>  6 prg_ever  0      never pregnant        FALSE  
#>  7 prg_ever  1      pregnant once or more FALSE  
#>  8 prg_ever  2      currently pregnant    FALSE  
#>  9 prg_ever  8      Don’t want to answer  TRUE   
#> 10 prg_ever  9      Don’t know            TRUE   
#> 11 prg_ever  -7     not applicable        TRUE   
#> 
#> 
#> $`dd_TOKYO_format_opal_tagged - ERROR WITH TAXO`
#> $`dd_TOKYO_format_opal_tagged - ERROR WITH TAXO`$Variables
#> # A tibble: 9 × 17
#>   name      `label:en`            `description:en`  valueType unit  other_format
#>   <chr>     <chr>                 <chr>             <chr>     <chr> <lgl>       
#> 1 part_id   id of the participant id of the partic… text      NA    NA          
#> 2 gndr      gndr                  gender of the pa… text      NA    NA          
#> 3 height    height                height of the pa… integer   cm    NA          
#> 4 weight_ms weight_ms             weight of the pa… integer   kg    NA          
#> 5 weight_dc weight_dc             weight of the pa… decimal   kg    NA          
#> 6 dob       dob                   date of birth of… date      years NA          
#> 7 prg_ever  prg_ever              whether the part… integer   NA    NA          
#> 8 empty     empty                 empty column      integer   NA    NA          
#> 9 opentext  opentext              open text         text      NA    NA          
#> # ℹ 11 more variables: `Mlstr_additional::Source` <chr>,
#> #   `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::Administrative_information` <chr>,
#> #   `Mlstr_area::Sociodemographic_economic_characteristics` <chr>,
#> #   `Mlstr_area::Cognitive_psychological_measures` <chr>,
#> #   `Mlstr_area::Physical_measures` <chr>, `Mlstr_area::Rep` <chr>,
#> #   `Mlstr_area::Non_pharmacological_interventions...14` <lgl>, …
#> 
#> $`dd_TOKYO_format_opal_tagged - ERROR WITH TAXO`$Categories
#> # A tibble: 9 × 4
#>   variable name   `label:en`            missing
#>   <chr>    <chr>  <chr>                   <dbl>
#> 1 gndr     Male   Male                        0
#> 2 gndr     Female Female                      0
#> 3 gndr     -77    Don’t want to answer        1
#> 4 prg_ever 0      never pregnant              0
#> 5 prg_ever 1      pregnant once or more       0
#> 6 prg_ever 2      currently pregnant          0
#> 7 prg_ever 8      Don’t want to answer        1
#> 8 prg_ever 9      Don’t know                  1
#> 9 prg_ever -7     not applicable              1
#> 
#> 
#> $dd_TOKYO_format_opal_tagged
#> $dd_TOKYO_format_opal_tagged$Variables
#> # A tibble: 9 × 13
#>   name      `label:en`            `description:en`  valueType unit  other_format
#>   <chr>     <chr>                 <chr>             <chr>     <chr> <lgl>       
#> 1 part_id   id of the participant id of the partic… text      NA    NA          
#> 2 gndr      gndr                  gender of the pa… text      NA    NA          
#> 3 height    height                height of the pa… integer   cm    NA          
#> 4 weight_ms weight_ms             weight of the pa… integer   kg    NA          
#> 5 weight_dc weight_dc             weight of the pa… decimal   kg    NA          
#> 6 dob       dob                   date of birth of… date      years NA          
#> 7 prg_ever  prg_ever              whether the part… integer   NA    NA          
#> 8 empty     empty                 empty column      integer   NA    NA          
#> 9 opentext  opentext              open text         text      NA    NA          
#> # ℹ 7 more variables: `Mlstr_additional::Source` <chr>,
#> #   `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::Administrative_information` <chr>,
#> #   `Mlstr_area::Sociodemographic_economic_characteristics` <chr>,
#> #   `Mlstr_area::Physical_measures` <chr>, `Mlstr_area::Reproduction` <chr>,
#> #   `Mlstr_area::Non_pharmacological_interventions` <chr>
#> 
#> $dd_TOKYO_format_opal_tagged$Categories
#> # A tibble: 9 × 4
#>   variable name   `label:en`            missing
#>   <chr>    <chr>  <chr>                   <dbl>
#> 1 gndr     Male   Male                        0
#> 2 gndr     Female Female                      0
#> 3 gndr     -77    Don’t want to answer        1
#> 4 prg_ever 0      never pregnant              0
#> 5 prg_ever 1      pregnant once or more       0
#> 6 prg_ever 2      currently pregnant          0
#> 7 prg_ever 8      Don’t want to answer        1
#> 8 prg_ever 9      Don’t know                  1
#> 9 prg_ever -7     not applicable              1
#> 
#> 
#> $dataset_MELBOURNE_1
#> # A tibble: 19 × 4
#>        id Gender   BMI   age
#>     <dbl>  <dbl> <dbl> <dbl>
#>  1 377943      2  22.1    52
#>  2 497013      1  18.5    49
#>  3 927676      1  24.6    43
#>  4 995667      2  15.7    59
#>  5  21829      2  18.6    40
#>  6 209432      1  15.9    47
#>  7 272983      2  28.7  -888
#>  8 580632      2  19.7    53
#>  9 304624      2  NA      35
#> 10 637551      1  21.3    40
#> 11 279817      1  17.2    41
#> 12 235415      2  21.7    34
#> 13 373673      2  14.7    48
#> 14 485098      2  24.4    43
#> 15 299427      1  NA    -888
#> 16 854073      1  24.2    41
#> 17 197666      2  17.7    33
#> 18 130327      2  19.2    57
#> 19 220050      2  18.8    50
#> 
#> $dataset_MELBOURNE_2
#> # A tibble: 19 × 3
#>        id smo_status prg_curr
#>     <dbl>      <dbl>    <dbl>
#>  1 377943          1        0
#>  2 497013          2      -77
#>  3 927676          3      -77
#>  4 995667        -77        1
#>  5  21829         NA        0
#>  6 209432          2      -77
#>  7 272983        -77        8
#>  8 580632          2        0
#>  9 304624          1        0
#> 10 637551          1      -77
#> 11 279817         NA      -77
#> 12 235415          3        1
#> 13 373673          2        1
#> 14 485098          1        9
#> 15 299427          2      -77
#> 16 854073          1      -77
#> 17 197666         NA        1
#> 18 130327          1        9
#> 19 220050          2        1
#> 
#> $dataset_PARIS
#> # A tibble: 24 × 7
#>    ID             SEX   BMI   AGE   SMO SMO_QTY PRG_EVER
#>    <chr>        <dbl> <dbl> <dbl> <dbl>   <dbl>    <dbl>
#>  1 Paris_687393     1  22.2    52     1      32        0
#>  2 Paris_585666     0  15.2    49     1       8       -8
#>  3 Paris_75802      0  22.7    43     1      48       -8
#>  4 Paris_412072     1  NA      59     1      11        0
#>  5 Paris_404333     1  26.2    40     1      18        1
#>  6 Paris_554985     0  16.0    47     1       7       -8
#>  7 Paris_714168     1  19.0    46     1      18       NA
#>  8 Paris_145477     1  16.9    53     0      -8        1
#>  9 Paris_202076     1  29.9    35     1      36        1
#> 10 Paris_847235     0  20.6    NA     0      -8       -8
#> # ℹ 14 more rows
#> 
#> $`dataset_TOKYO - ERROR WITH DATA`
#> # A tibble: 50 × 9
#>    part_id gndr   height weight_ms weight_dc dob        prg_ever empty opentext 
#>    <chr>   <chr>   <dbl>     <dbl>     <dbl> <chr>         <dbl> <lgl> <chr>    
#>  1 ID001   Male      191        -7      NA   3/22/1990        -7 NA    All chil…
#>  2 ID002   Female    191         0      65.3 8/15/2001         0 NA    grow up,…
#>  3 ID003   Female    191         8      45   12/17/1996        8 NA    rather d…
#>  4 ID004   Female    191         1      NA   6/13/1990         1 NA    flower a…
#>  5 ID005   Female    191         8      45   12/17/1996        8 NA    rather d…
#>  6 ID006   -77       191        -8      NA   3/31/1981        -8 NA    cried, O…
#>  7 ID007   Female    191         9      58   4/19/1988         9 NA    that pas…
#>  8 ID008   NA         NA        NA      NA   NA               NA NA    NA       
#>  9 ID009   -77       191        -7      NA   3/14/1976        -7 NA    beginnin…
#> 10 ID010   Male      191        -7      65.3 10/19/1993       -7 NA    All chil…
#> # ℹ 40 more rows
#> 
#> $dataset_TOKYO
#> # A tibble: 50 × 9
#>    part_id gndr   height weight_ms weight_dc dob        prg_ever empty opentext 
#>    <chr>   <chr>   <dbl>     <dbl>     <dbl> <chr>         <dbl> <lgl> <chr>    
#>  1 ID001   Male      191        63        NA 3/22/1990        -7 NA    All chil…
#>  2 ID002   Female    176        NA        65 8/15/2001         0 NA    grow up,…
#>  3 ID003   Female    154        NA       121 12/17/1996        2 NA    years ol…
#>  4 ID004   Female    167       -88        NA 6/13/1990         1 NA    flower a…
#>  5 ID005   Female    185        NA        45 12/17/1996        8 NA    rather d…
#>  6 ID006   -77       171        57        NA 3/31/1981        -7 NA    cried, O…
#>  7 ID007   Female    185        NA        58 4/19/1988         9 NA    that pas…
#>  8 ID008   Female    171        NA        59 NA                2 NA    that she…
#>  9 ID009   -77       169        52        NA 3/14/1976        -7 NA    beginnin…
#> 10 ID010   Male      179        NA        62 10/19/1993       -7 NA    All chil…
#> # ℹ 40 more rows
#> 
#> $taxonomy_PARIS
#> # A tibble: 15 × 7
#>    index_taxonomy index_vocabulary index_term taxonomy_opal  taxonomy vocabulary
#>             <dbl>            <dbl>      <dbl> <chr>          <chr>    <chr>     
#>  1              1                1          1 description::… descrip… type      
#>  2              1                1          2 description::… descrip… type      
#>  3              1                1          3 description::… descrip… type      
#>  4              1                2          1 description::… descrip… level     
#>  5              1                2          2 description::… descrip… level     
#>  6              1                2          3 description::… descrip… level     
#>  7              2                1          1 auth::importa… auth     importance
#>  8              2                1          2 auth::importa… auth     importance
#>  9              3                1          1 relevance::ac… relevan… accuracy  
#> 10              3                1          2 relevance::ac… relevan… accuracy  
#> 11              3                1          3 relevance::ac… relevan… accuracy  
#> 12              3                1          4 relevance::ac… relevan… accuracy  
#> 13              3                1          5 relevance::ac… relevan… accuracy  
#> 14              4                1          1 Mlstr_additio… Mlstr_a… Source    
#> 15              4                1          2 Mlstr_additio… Mlstr_a… Source    
#> # ℹ 1 more variable: term <chr>
#> 
#> $taxonomy_opal
#> # A tibble: 1,073 × 6
#>    index_taxonomy index_vocabulary index_term taxonomy         vocabulary term  
#>             <int>            <int>      <int> <chr>            <chr>      <chr> 
#>  1              1                1          1 Mlstr_additional Source     Quest…
#>  2              1                1          2 Mlstr_additional Source     Physi…
#>  3              1                1          3 Mlstr_additional Source     Cogni…
#>  4              1                1          4 Mlstr_additional Source     Biosa…
#>  5              1                1          5 Mlstr_additional Source     Gover…
#>  6              1                1          6 Mlstr_additional Source     Other 
#>  7              1                2          1 Mlstr_additional Target     Parti…
#>  8              1                2          2 Mlstr_additional Target     Child…
#>  9              1                2          3 Mlstr_additional Target     Paren…
#> 10              1                2          4 Mlstr_additional Target     Grand…
#> # ℹ 1,063 more rows
#> 
#> $taxonomy_opal_mlstr
#> # A tibble: 1,073 × 11
#>    index_taxonomy index_vocabulary index_term taxonomy         vocabulary term  
#>  *          <int>            <int>      <int> <chr>            <chr>      <chr> 
#>  1              1                1          1 Mlstr_additional Source     Quest…
#>  2              1                1          2 Mlstr_additional Source     Physi…
#>  3              1                1          3 Mlstr_additional Source     Cogni…
#>  4              1                1          4 Mlstr_additional Source     Biosa…
#>  5              1                1          5 Mlstr_additional Source     Gover…
#>  6              1                1          6 Mlstr_additional Source     Other 
#>  7              1                2          1 Mlstr_additional Target     Parti…
#>  8              1                2          2 Mlstr_additional Target     Child…
#>  9              1                2          3 Mlstr_additional Target     Paren…
#> 10              1                2          4 Mlstr_additional Target     Grand…
#> # ℹ 1,063 more rows
#> # ℹ 5 more variables: vocabulary_short <chr>, index_term_scale <int>,
#> #   taxonomy_scale <chr>, vocabulary_scale <chr>, term_scale <chr>
#>