Validate and coerce an object to taxonomy format

Confirms that the input object is a valid taxonomy and returns it as a taxonomy with the appropriate madshapR::class attribute. This function mainly helps validate input within other functions of the package but could be used to check if a taxonomy is valid.

as_taxonomy(object)

Arguments

object

A potential taxonomy to be coerced.

Value

A tibble identifying a taxonomy (generally generated from Opal taxonomy).

Details

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# use DEMO_files provided by the package

###### Example 1: this function is compatible with taxonomy_opal_get()
library(opalr)
opal <- 
  opal.login('administrator','password',
    url = 'https://opal-demo.maelstrom-research.org/')
opal_taxo <- taxonomy_opal_mlstr_get(opal)
as_taxonomy(opal_taxo)

###### Example 2: you can create your own taxonomy
print(DEMO_files$taxonomy_PARIS)
as_taxonomy(DEMO_files$taxonomy_PARIS)

}
#> Loading required package: httr
#> # A tibble: 15 × 7
#>    index_taxonomy index_vocabulary index_term taxonomy_opal  taxonomy vocabulary
#>             <dbl>            <dbl>      <dbl> <chr>          <chr>    <chr>     
#>  1              1                1          1 description::… descrip… type      
#>  2              1                1          2 description::… descrip… type      
#>  3              1                1          3 description::… descrip… type      
#>  4              1                2          1 description::… descrip… level     
#>  5              1                2          2 description::… descrip… level     
#>  6              1                2          3 description::… descrip… level     
#>  7              2                1          1 auth::importa… auth     importance
#>  8              2                1          2 auth::importa… auth     importance
#>  9              3                1          1 relevance::ac… relevan… accuracy  
#> 10              3                1          2 relevance::ac… relevan… accuracy  
#> 11              3                1          3 relevance::ac… relevan… accuracy  
#> 12              3                1          4 relevance::ac… relevan… accuracy  
#> 13              3                1          5 relevance::ac… relevan… accuracy  
#> 14              4                1          1 Mlstr_additio… Mlstr_a… Source    
#> 15              4                1          2 Mlstr_additio… Mlstr_a… Source    
#> # ℹ 1 more variable: term <chr>
#> # A tibble: 15 × 7
#>    index_taxonomy index_vocabulary index_term taxonomy_opal  taxonomy vocabulary
#>  *          <dbl>            <dbl>      <dbl> <chr>          <chr>    <chr>     
#>  1              1                1          1 description::… descrip… type      
#>  2              1                1          2 description::… descrip… type      
#>  3              1                1          3 description::… descrip… type      
#>  4              1                2          1 description::… descrip… level     
#>  5              1                2          2 description::… descrip… level     
#>  6              1                2          3 description::… descrip… level     
#>  7              2                1          1 auth::importa… auth     importance
#>  8              2                1          2 auth::importa… auth     importance
#>  9              3                1          1 relevance::ac… relevan… accuracy  
#> 10              3                1          2 relevance::ac… relevan… accuracy  
#> 11              3                1          3 relevance::ac… relevan… accuracy  
#> 12              3                1          4 relevance::ac… relevan… accuracy  
#> 13              3                1          5 relevance::ac… relevan… accuracy  
#> 14              4                1          1 Mlstr_additio… Mlstr_a… Source    
#> 15              4                1          2 Mlstr_additio… Mlstr_a… Source    
#> # ℹ 1 more variable: term <chr>