Apply a data dictionary to a dataset

Applies a data dictionary to a data structure, creating a labelled dataset. All previous attributes will be preserved. For factors, the attribute 'levels' will be transformed into attribute 'labels' and values will be recast into appropriate datatypes.

data_dict_apply(dataset, data_dict = NULL)

Arguments

dataset

A data frame identifying the input dataset observations associated to its data dictionary.

data_dict

A list of data frame(s) representing meta data of an associated dataset. Automatically generated if not provided.

Value

A data frame identifying the dataset with the data dictionary applied to each variable as attributes.

Details

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

See also

Examples

{

# use madshapR_DEMO provided by the package

dataset <- madshapR_DEMO$dataset_MELBOURNE
data_dict <- as_data_dict_mlstr(madshapR_DEMO$data_dict_MELBOURNE)
data_dict_apply(dataset, data_dict)

}
#> # A tibble: 19 × 6
#>        id Gender            BMI age                         smo_status prg_curr 
#>  *  <dbl> <dbl+lbl>       <dbl> <dbl+lbl>                   <dbl+lbl>  <dbl+lbl>
#>  1 377943 2 [Female]        221   52                          1 [neve…   0 [not…
#>  2 497013 1 [Male]   1850655594   49                          2 [curr… -77 [not…
#>  3 927676 1 [Male]   2457679588   43                          3 [form… -77 [not…
#>  4 995667 2 [Female] 1571539833   59                        -77 [don'…   1 [cur…
#>  5  21829 2 [Female] 1855378065   40                         NA          0 [not…
#>  6 209432 1 [Male]   1593177939   47                          2 [curr… -77 [not…
#>  7 272983 2 [Female] 2865839945 -888 [don't want to answer] -77 [don'…   8 [Don…
#>  8 580632 2 [Female] 1973602923   53                          2 [curr…   0 [not…
#>  9 304624 2 [Female]         NA   35                          1 [neve…   0 [not…
#> 10 637551 1 [Male]   2126602756   40                          1 [neve… -77 [not…
#> 11 279817 1 [Male]   1719871484   41                         NA        -77 [not…
#> 12 235415 2 [Female] 2169671091   34                          3 [form…   1 [cur…
#> 13 373673 2 [Female] 1474286366   48                          2 [curr…   1 [cur…
#> 14 485098 2 [Female] 2437040385   43                          1 [neve…   9 [Don…
#> 15 299427 1 [Male]           NA -888 [don't want to answer]   2 [curr… -77 [not…
#> 16 854073 1 [Male]   2420491247   41                          1 [neve… -77 [not…
#> 17 197666 2 [Female] 1773788018   33                         NA          1 [cur…
#> 18 130327 2 [Female] 1915254927   57                          1 [neve…   9 [Don…
#> 19 220050 2 [Female] 1876765116   50                          2 [curr…   1 [cur…