Generate a quality assessment report of a data dictionary

Assesses the content and structure of a data dictionary and reports potential issues to facilitate the assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. This report is compatible with Excel and can be exported as an Excel spreadsheet.

data_dict_evaluate(data_dict, taxonomy = NULL, as_data_dict_mlstr = TRUE)

Arguments

data_dict

A list of data frame(s) representing meta data to be evaluated.

taxonomy

An optional data frame identifying a variable classification schema.

as_data_dict_mlstr

Whether the output data dictionary has a simple data dictionary structure or not (meaning has a Maelstrom data dictionary structure, compatible with Maelstrom Research ecosystem, including Opal). TRUE by default.

Value

A list of data frames of report for one data dictionary.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

A taxonomy is a classification schema that can be defined for variable attributes. A taxonomy is usually extracted from an Opal environment, and a taxonomy object is a data frame that must contain at least the columns taxonomy, vocabulary, and terms. Additional details about Opal taxonomies are available online.

The object may be specifically formatted to be compatible with additional Maelstrom Research software, in particular Opal environments.

Examples

{

# use madshapR_DEMO provided by the package

data_dict <- madshapR_DEMO$`data_dict_TOKYO - errors`
data_dict_evaluate(data_dict)

}
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> $`Data dictionary summary`
#> # A tibble: 14 × 11
#>    index name    `label:en` valueType `Categories::table` `Categories::label:en`
#>    <int> <chr>   <chr>      <chr>     <chr>               <chr>                 
#>  1     1 part_id id of the… text       NA                  NA                   
#>  2     2 gndr    gndr       boolean   "Male = DEMO ; \nF… "Male = Male ; \nFema…
#>  3     3 height  height     decimal   "SKIP PATTERN = DE… "SKIP PATTERN = SKIP …
#>  4     4 weight… weight_ms  NA        "-88 = DEMO ; \n-9… "-88 = Don’t want to …
#>  5     5 weight… weight_dc  integer    NA                  NA                   
#>  6     6 weight… weight_dc  integr     NA                  NA                   
#>  7     7 dob     NA         date       NA                  NA                   
#>  8     8 prg_evr prg_ever   integer    NA                  NA                   
#>  9     9 empty   empty      integer    NA                  NA                   
#> 10    10 opente… opentext   character  NA                  NA                   
#> 11    11 opente… NA         NA         NA                  NA                   
#> 12    12 NA      no name    integer   "6 = DEMO"          "6 = Inaccurate"      
#> 13    NA prg_ev… NA         NA        "0 = DEMO ; \n1 = … "0 = never pregnant ;…
#> 14    NA weight… NA         NA        "-99 = DEMO"        "-99 = Don’t want to …
#> # ℹ 5 more variables: `Categories::missing` <chr>, table <chr>,
#> #   `description:en` <chr>, unit <chr>, `copy units` <chr>
#> 
#> $`Data dictionary assessment`
#> # A tibble: 18 × 6
#>    sheet      col_name   name_var       Quality assessment co…¹ value suggestion
#>    <chr>      <chr>      <chr>          <chr>                   <chr> <chr>     
#>  1 Variables  copy units NA             [INFO] - Possible dupl… unit… NA        
#>  2 Variables  label:en   dob            [ERR] - The column `la… NA    NA        
#>  3 Variables  label:en   opentext       [ERR] - The column `la… NA    NA        
#>  4 Variables  name       opentext       [ERR] - duplicated var… NA    NA        
#>  5 Variables  name       row number: 12 [ERR] - missing variab… NA    NA        
#>  6 Variables  name       weight dc      [INFO] - Incompatible … NA    NA        
#>  7 Variables  unit       NA             [INFO] - Possible dupl… unit… NA        
#>  8 Variables  valueType  gndr           [ERR]  - valueType con… bool… text      
#>  9 Variables  valueType  height         [ERR]  - valueType con… deci… text      
#> 10 Variables  valueType  opentext       [ERR] - Incompatible v… char… NA        
#> 11 Variables  valueType  weight_dc      [ERR] - Incompatible v… inte… NA        
#> 12 Categories label:en   1              [ERR] - The column `la… NA    NA        
#> 13 Categories missing    prg_ever       [ERR] - incompatible v… FLASE NA        
#> 14 Categories variable   prg_ever       [ERR] - Categories not… NA    NA        
#> 15 Categories variable   prg_ever       [ERR] - In 'name', the… row … NA        
#> 16 Categories variable   prg_ever       [ERR] - Category names… -7    NA        
#> 17 Categories variable   weight_sm      [ERR] - Categories not… NA    NA        
#> 18 Categories variable   NA             [ERR] - In 'variable',… row … NA        
#> # ℹ abbreviated name: ¹​`Quality assessment comment`
#>