Subset data dictionary by row values

Subsets either or both the 'Variables' and 'Categories' elements of a data dictionary. Rows are conserved if their values satisfy the condition. This is a wrapper function analogous to dplyr::filter().

data_dict_filter(
  data_dict,
  filter_var = NULL,
  filter_cat = NULL,
  filter_all = NULL
)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed.

filter_var

Expressions that are defined in the element 'Variables' in the data dictionary.

filter_cat

Expressions that are defined in the element 'Categories' in the data dictionary.

filter_all

Expressions that are defined both in the 'Categories' and 'Variables' in the data dictionary.

Value

A list of tibble(s) identifying a workable data dictionary structure.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

See also

Examples

{

# use DEMO_files provided by the package

# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a 
# data dictionary per say, but can be used as a structure which can be 
# shaped into a data dictionary.
library(dplyr)

data_dict_list <- list()
data_dict_1 <-
  DEMO_files$dd_MELBOURNE_1_format_maelstrom %>%
  lapply(function(x){x %>% mutate(table = "MELBOURNE_1")})
data_dict_2 <- DEMO_files$dd_MELBOURNE_2_format_maelstrom %>%
  lapply(function(x){x %>% mutate(table = "MELBOURNE_2")})

data_dict_list <-
  list(Variables = bind_rows(data_dict_1$Variables,data_dict_2$Variables),
       Categories = bind_rows(data_dict_1$Categories,data_dict_2$Categories))

###### Example 1 search and filter through a column in 'Variables' element
data_dict_filter(data_dict_list,filter_var = "valueType == 'integer'")

###### Example 2 search and filter through a column in 'Categories' element
data_dict_filter(data_dict_list,filter_cat = "missing == TRUE")

###### Example 3 search and filter through a column in 'Variables' element.
# The column must exist in both 'Variables' and 'Categories' and have the 
# same meaning
data_dict_filter(data_dict_list,filter_all = "table == 'MELBOURNE_1'")

}
#> $Variables
#> # A tibble: 4 × 8
#>   index name   `label:en` `description:en`    valueType unit  other_format table
#>   <dbl> <chr>  <chr>      <chr>               <chr>     <chr> <lgl>        <chr>
#> 1     1 id     id         id of the particip… text      NA    NA           MELB…
#> 2     2 Gender Gender     Gender              integer   NA    NA           MELB…
#> 3     3 BMI    BMI        Body Mass Index     decimal   kg/m… NA           MELB…
#> 4     4 age    age        Age of Participant  integer   years NA           MELB…
#> 
#> $Categories
#> # A tibble: 3 × 5
#>   variable  name `label:en`           missing table      
#>   <chr>    <dbl> <chr>                <lgl>   <chr>      
#> 1 age       -888 don't want to answer TRUE    MELBOURNE_1
#> 2 Gender       1 Male                 FALSE   MELBOURNE_1
#> 3 Gender       2 Female               FALSE   MELBOURNE_1
#>