Group listed data dictionaries by specified column names

Groups the data dictionary element(s) by the groups defined by the query. This function groups both the 'Variables' and 'Categories' elements (if the group exists under the same definition in in both). This function is analagous to running dplyr::group_by(). Each element is named using the group values. data_dict_ungroup() reverses the effect.

data_dict_group_by(data_dict, col)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed.

col

variable to group by.

Value

A list of tibble(s) identifying a workable data dictionary structure.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

See also

Examples

{

# use DEMO_files provided by the package
# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a 
# data dictionary per say, but can be used as a structure which can be shaped 
# into a data dictionary.

library(dplyr)

data_dict_list <- 
  DEMO_files[c('dd_MELBOURNE_1_format_maelstrom',
               'dd_MELBOURNE_2_format_maelstrom')] %>%
data_dict_list_nest(name_group = 'table')
 
data_dict_group_by(data_dict_list,col = "table")
}
#> $Variables
#> # A tibble: 7 × 8
#> # Groups:   table [2]
#>   table     index name  `label:en` `description:en` valueType unit  other_format
#>   <chr>     <chr> <chr> <chr>      <chr>            <chr>     <chr> <chr>       
#> 1 dd_MELBO… 1     id    id         id of the parti… text      NA    NA          
#> 2 dd_MELBO… 2     Gend… Gender     Gender           integer   NA    NA          
#> 3 dd_MELBO… 3     BMI   BMI        Body Mass Index  decimal   kg/m… NA          
#> 4 dd_MELBO… 4     age   age        Age of Particip… integer   years NA          
#> 5 dd_MELBO… 1     id    id         id of the parti… text      NA    NA          
#> 6 dd_MELBO… 2     smo_… smo_curr   Whether the par… integer   NA    NA          
#> 7 dd_MELBO… 3     prg_… prg_curr   Are you current… integer   NA    NA          
#> 
#> $Categories
#> # A tibble: 12 × 5
#> # Groups:   table [2]
#>    table                           variable   name  `label:en`           missing
#>    <chr>                           <chr>      <chr> <chr>                <chr>  
#>  1 dd_MELBOURNE_1_format_maelstrom age        -888  don't want to answer TRUE   
#>  2 dd_MELBOURNE_1_format_maelstrom Gender     1     Male                 FALSE  
#>  3 dd_MELBOURNE_1_format_maelstrom Gender     2     Female               FALSE  
#>  4 dd_MELBOURNE_2_format_maelstrom prg_curr   0     not currently pregn… FALSE  
#>  5 dd_MELBOURNE_2_format_maelstrom prg_curr   1     currently pregnant   FALSE  
#>  6 dd_MELBOURNE_2_format_maelstrom prg_curr   8     Don’t want to answer TRUE   
#>  7 dd_MELBOURNE_2_format_maelstrom prg_curr   9     Don’t know           TRUE   
#>  8 dd_MELBOURNE_2_format_maelstrom prg_curr   -77   not applicable       TRUE   
#>  9 dd_MELBOURNE_2_format_maelstrom smo_status 1     never smoked         FALSE  
#> 10 dd_MELBOURNE_2_format_maelstrom smo_status 2     current smoker       FALSE  
#> 11 dd_MELBOURNE_2_format_maelstrom smo_status 3     former smoker        FALSE  
#> 12 dd_MELBOURNE_2_format_maelstrom smo_status -77   don't want to answer TRUE   
#>