Split grouped data dictionaries into a named list

Divides data dictionary element(s) into the groups defined by the query. This function divides both the 'Variables' and 'Categories' elements (if the group exists under the same definition in in both) into a list of data dictionaries, each with the rows of the associated group and all the original columns, including grouping variables. This function is analogous to running dplyr::group_by() and dplyr::group_split(). Each element is named using the group values. data_dict_list_nest() reverses the effect.

data_dict_group_split(data_dict, ...)

Arguments

data_dict

A list of data frame(s) representing meta data to be transformed.

...

Column in the data dictionary to split it by. If not provided, the splitting will be done on the grouping element of a grouped data dictionary.

Value

A list of data frame(s) identifying a list of workable data dictionary structure.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

See also

Examples

{

# use madshapR_DEMO provided by the package
library(dplyr)

# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a 
# data dictionary per say, but can be used as a structure which can be 
# shaped into a data dictionary.

data_dict_list <- list(
  data_dict_1 <- madshapR_DEMO$data_dict_TOKYO ,
  data_dict_2 <- madshapR_DEMO$data_dict_MELBOURNE)
names(data_dict_list) = c("dataset_TOKYO","dataset_MELBOURNE")

data_dict_nest <- 
  data_dict_list_nest(data_dict_list, name_group = 'table') %>%
  data_dict_group_by(col = "table")

 data_dict_group_split(data_dict_nest,col = "table")
 
}
#> $dataset_MELBOURNE
#> $dataset_MELBOURNE$Variables
#> # A tibble: 6 × 7
#>   table             index name       `label:en` `description:en` valueType unit 
#>   <chr>             <chr> <chr>      <chr>      <chr>            <chr>     <chr>
#> 1 dataset_MELBOURNE 1     id         id         id of the parti… text      NA   
#> 2 dataset_MELBOURNE 2     Gender     Gender     Gender           integer   NA   
#> 3 dataset_MELBOURNE 3     BMI        BMI        Body Mass Index  decimal   kg/m…
#> 4 dataset_MELBOURNE 4     age        age        Age of Particip… integer   years
#> 5 dataset_MELBOURNE 5     smo_status smo_curr   Whether the par… integer   NA   
#> 6 dataset_MELBOURNE 6     prg_curr   prg_curr   Are you current… integer   NA   
#> 
#> $dataset_MELBOURNE$Categories
#> # A tibble: 12 × 5
#>    table             variable   name  `label:en`             missing
#>    <chr>             <chr>      <chr> <chr>                  <chr>  
#>  1 dataset_MELBOURNE age        -888  don't want to answer   TRUE   
#>  2 dataset_MELBOURNE Gender     1     Male                   FALSE  
#>  3 dataset_MELBOURNE Gender     2     Female                 FALSE  
#>  4 dataset_MELBOURNE prg_curr   0     not currently pregnant FALSE  
#>  5 dataset_MELBOURNE prg_curr   1     currently pregnant     FALSE  
#>  6 dataset_MELBOURNE prg_curr   8     Don’t want to answer   TRUE   
#>  7 dataset_MELBOURNE prg_curr   9     Don’t know             TRUE   
#>  8 dataset_MELBOURNE prg_curr   -77   not applicable         TRUE   
#>  9 dataset_MELBOURNE smo_status 1     never smoked           FALSE  
#> 10 dataset_MELBOURNE smo_status 2     current smoker         FALSE  
#> 11 dataset_MELBOURNE smo_status 3     former smoker          FALSE  
#> 12 dataset_MELBOURNE smo_status -77   don't want to answer   TRUE   
#> 
#> 
#> $dataset_TOKYO
#> $dataset_TOKYO$Variables
#> # A tibble: 9 × 7
#>   table         index name      `label:en`      `description:en` valueType unit 
#>   <chr>         <chr> <chr>     <chr>           <chr>            <chr>     <chr>
#> 1 dataset_TOKYO 1     part_id   id of the part… id of the parti… text      NA   
#> 2 dataset_TOKYO 2     gndr      gndr            gender of the p… text      NA   
#> 3 dataset_TOKYO 3     height    height          height of the p… integer   cm   
#> 4 dataset_TOKYO 4     weight_ms weight_ms       weight of the p… integer   kg   
#> 5 dataset_TOKYO 5     weight_dc weight_dc       weight of the p… decimal   kg   
#> 6 dataset_TOKYO 6     dob       dob             date of birth o… date      years
#> 7 dataset_TOKYO 7     prg_ever  prg_ever        whether the par… integer   NA   
#> 8 dataset_TOKYO 8     empty     empty           empty column     integer   NA   
#> 9 dataset_TOKYO 9     opentext  opentext        open text        text      NA   
#> 
#> $dataset_TOKYO$Categories
#> # A tibble: 11 × 5
#>    table         variable  name   `label:en`            missing
#>    <chr>         <chr>     <chr>  <chr>                 <chr>  
#>  1 dataset_TOKYO gndr      Male   Male                  FALSE  
#>  2 dataset_TOKYO gndr      Female Female                FALSE  
#>  3 dataset_TOKYO gndr      -77    Don’t want to answer  TRUE   
#>  4 dataset_TOKYO weight_ms -88    Don’t want to answer  TRUE   
#>  5 dataset_TOKYO weight_ms -99    Don’t know            TRUE   
#>  6 dataset_TOKYO prg_ever  0      never pregnant        FALSE  
#>  7 dataset_TOKYO prg_ever  1      pregnant once or more FALSE  
#>  8 dataset_TOKYO prg_ever  2      currently pregnant    FALSE  
#>  9 dataset_TOKYO prg_ever  8      Don’t want to answer  TRUE   
#> 10 dataset_TOKYO prg_ever  9      Don’t know            TRUE   
#> 11 dataset_TOKYO prg_ever  -7     not applicable        TRUE   
#> 
#>