Bind listed data dictionaries

Binds a list of data dictionaries into one data dictionary. This is a wrapper function analogous to dplyr::bind_rows().

data_dict_list_nest(data_dict_list, name_group = NULL)

Arguments

data_dict_list

A list of data frame(s) representing meta data to be transformed.

name_group

A character string of one column in the dataset that can be taken as a grouping column.

Value

A list of data frame(s) identifying a workable data dictionary structure.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

See also

Examples

{

# use madshapR_DEMO provided by the package
# Create a list of data dictionaries where the column 'table' is added to 
# refer to the associated dataset. The object created is not a 
# data dictionary per say, but can be used as a structure which can be 
# shaped into a data dictionary.

data_dict_list <- list(
  data_dict_1 <- madshapR_DEMO$data_dict_TOKYO ,
  data_dict_2 <- madshapR_DEMO$data_dict_MELBOURNE)
names(data_dict_list) = c("dataset_TOKYO","dataset_MELBOURNE")

data_dict_list_nest(data_dict_list, name_group = 'table')

}
#> $Variables
#> # A tibble: 15 × 7
#>    table             index name      `label:en` `description:en` valueType unit 
#>    <chr>             <chr> <chr>     <chr>      <chr>            <chr>     <chr>
#>  1 dataset_TOKYO     1     part_id   id of the… id of the parti… text      NA   
#>  2 dataset_TOKYO     2     gndr      gndr       gender of the p… text      NA   
#>  3 dataset_TOKYO     3     height    height     height of the p… integer   cm   
#>  4 dataset_TOKYO     4     weight_ms weight_ms  weight of the p… integer   kg   
#>  5 dataset_TOKYO     5     weight_dc weight_dc  weight of the p… decimal   kg   
#>  6 dataset_TOKYO     6     dob       dob        date of birth o… date      years
#>  7 dataset_TOKYO     7     prg_ever  prg_ever   whether the par… integer   NA   
#>  8 dataset_TOKYO     8     empty     empty      empty column     integer   NA   
#>  9 dataset_TOKYO     9     opentext  opentext   open text        text      NA   
#> 10 dataset_MELBOURNE 1     id        id         id of the parti… text      NA   
#> 11 dataset_MELBOURNE 2     Gender    Gender     Gender           integer   NA   
#> 12 dataset_MELBOURNE 3     BMI       BMI        Body Mass Index  decimal   kg/m…
#> 13 dataset_MELBOURNE 4     age       age        Age of Particip… integer   years
#> 14 dataset_MELBOURNE 5     smo_stat… smo_curr   Whether the par… integer   NA   
#> 15 dataset_MELBOURNE 6     prg_curr  prg_curr   Are you current… integer   NA   
#> 
#> $Categories
#> # A tibble: 23 × 5
#>    table         variable  name   `label:en`            missing
#>    <chr>         <chr>     <chr>  <chr>                 <chr>  
#>  1 dataset_TOKYO gndr      Male   Male                  FALSE  
#>  2 dataset_TOKYO gndr      Female Female                FALSE  
#>  3 dataset_TOKYO gndr      -77    Don’t want to answer  TRUE   
#>  4 dataset_TOKYO weight_ms -88    Don’t want to answer  TRUE   
#>  5 dataset_TOKYO weight_ms -99    Don’t know            TRUE   
#>  6 dataset_TOKYO prg_ever  0      never pregnant        FALSE  
#>  7 dataset_TOKYO prg_ever  1      pregnant once or more FALSE  
#>  8 dataset_TOKYO prg_ever  2      currently pregnant    FALSE  
#>  9 dataset_TOKYO prg_ever  8      Don’t want to answer  TRUE   
#> 10 dataset_TOKYO prg_ever  9      Don’t know            TRUE   
#> # ℹ 13 more rows
#>