Inner join between a dataset and its associated data dictionary

Performs an inner join between a dataset and its associated data dictionary, keeping only variables present in both. This function returns the matched dataset rows, the matched data dictionary rows, or both, in a list.

data_dict_match_dataset(
  dataset,
  data_dict,
  data_dict_apply = FALSE,
  output = c("dataset", "data_dict")
)

Arguments

dataset

A data frame identifying the input dataset observations.

data_dict

A list of data frame(s) representing meta data associated to a dataset.

data_dict_apply

whether to apply the data dictionary to its dataset. The resulting data frame will have for each column its associated meta data as attributes. FALSE by default.

output

A vector of character string which indicates if the function returns a dataset ('dataset'), data dictionary ('data_dict') of both. Default is c('dataset','data_dict').

Value

Either a data frame, identifying the dataset, or a list of data frame(s) identifying a data dictionary. Returns both in a list by default.

Details

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

Examples

{

# use madshapR_DEMO provided by the package
library(dplyr)
dataset <- madshapR_DEMO$dataset_MELBOURNE %>% select(-1)
data_dict <- madshapR_DEMO$data_dict_MELBOURNE
data_dict_match_dataset(dataset, data_dict)

}
#> $dataset
#> # A tibble: 19 × 5
#>    Gender        BMI   age smo_status prg_curr
#>     <dbl>      <dbl> <dbl>      <dbl>    <dbl>
#>  1      2        221    52          1        0
#>  2      1 1850655594    49          2      -77
#>  3      1 2457679588    43          3      -77
#>  4      2 1571539833    59        -77        1
#>  5      2 1855378065    40         NA        0
#>  6      1 1593177939    47          2      -77
#>  7      2 2865839945  -888        -77        8
#>  8      2 1973602923    53          2        0
#>  9      2         NA    35          1        0
#> 10      1 2126602756    40          1      -77
#> 11      1 1719871484    41         NA      -77
#> 12      2 2169671091    34          3        1
#> 13      2 1474286366    48          2        1
#> 14      2 2437040385    43          1        9
#> 15      1         NA  -888          2      -77
#> 16      1 2420491247    41          1      -77
#> 17      2 1773788018    33         NA        1
#> 18      2 1915254927    57          1        9
#> 19      2 1876765116    50          2        1
#> 
#> $data_dict
#> $data_dict$Variables
#> # A tibble: 5 × 6
#>   index name       `label:en` `description:en`                   valueType unit 
#>   <dbl> <chr>      <chr>      <chr>                              <chr>     <chr>
#> 1     2 Gender     Gender     Gender                             integer   NA   
#> 2     3 BMI        BMI        Body Mass Index                    decimal   kg/m…
#> 3     4 age        age        Age of Participant                 integer   years
#> 4     5 smo_status smo_curr   Whether the participant is a curr… integer   NA   
#> 5     6 prg_curr   prg_curr   Are you currently pregnant ?       integer   NA   
#> 
#> $data_dict$Categories
#> # A tibble: 12 × 4
#>    variable    name `label:en`             missing
#>    <chr>      <dbl> <chr>                  <lgl>  
#>  1 age         -888 don't want to answer   TRUE   
#>  2 Gender         1 Male                   FALSE  
#>  3 Gender         2 Female                 FALSE  
#>  4 prg_curr       0 not currently pregnant FALSE  
#>  5 prg_curr       1 currently pregnant     FALSE  
#>  6 prg_curr       8 Don’t want to answer   TRUE   
#>  7 prg_curr       9 Don’t know             TRUE   
#>  8 prg_curr     -77 not applicable         TRUE   
#>  9 smo_status     1 never smoked           FALSE  
#> 10 smo_status     2 current smoker         FALSE  
#> 11 smo_status     3 former smoker          FALSE  
#> 12 smo_status   -77 don't want to answer   TRUE   
#> 
#>