Turn column(s) in a data dictionary component into their longer equivalent

Transforms column(s) in a data dictionary component into their longer equivalent. If any Opal taxonomy is provided, the correspondent columns in the data dictionary will be converted to a format that contains fewer columns than an Opal format to work with. This operation is equivalent to perform a tidyr::pivot_longer() to these columns following the taxonomy structure provided. The data dictionary provided must respect certain constrains such as name uniqueness across the variables names.

data_dict_pivot_longer(data_dict, taxonomy = NULL)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed.

taxonomy

A data-frame or data-frame extension (e.g. a tibble), identifying the scheme used for variables classification as a tibble.

Value

A list of tibble(s) identifying a data dictionary.

Details

A data dictionary-like structure must be a list of at least one or two data-frame or data-frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A taxonomy must be a data-frame or data-frame extension (e.g. a tibble). The taxonomy must be compatible with (and generally extracted from) an Opal environment, and must contain at least 'taxonomy', 'vocabulary' and 'terms' to work with some specific functions. In addition, the taxonomy may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# use DEMO_files provided by the package

taxonomy <- DEMO_files$taxonomy_opal_mlstr
data_dict <- DEMO_files$dd_TOKYO_format_opal_tagged
data_dict_pivot_longer(data_dict,taxonomy)

}
#> $Variables
#> # A tibble: 9 × 14
#>   name      `label:en`            `description:en`  valueType unit  other_format
#>   <chr>     <chr>                 <chr>             <chr>     <chr> <lgl>       
#> 1 part_id   id of the participant id of the partic… text      NA    NA          
#> 2 gndr      gndr                  gender of the pa… text      NA    NA          
#> 3 height    height                height of the pa… integer   cm    NA          
#> 4 weight_ms weight_ms             weight of the pa… integer   kg    NA          
#> 5 weight_dc weight_dc             weight of the pa… decimal   kg    NA          
#> 6 dob       dob                   date of birth of… date      years NA          
#> 7 prg_ever  prg_ever              whether the part… integer   NA    NA          
#> 8 empty     empty                 empty column      integer   NA    NA          
#> 9 opentext  opentext              open text         text      NA    NA          
#> # ℹ 8 more variables: `Mlstr_additional::1` <chr>,
#> #   `Mlstr_additional::1.term` <chr>, `Mlstr_additional::2` <chr>,
#> #   `Mlstr_additional::2.term` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::2` <chr>,
#> #   `Mlstr_area::2.term` <chr>
#> 
#> $Categories
#> # A tibble: 9 × 4
#>   variable name   `label:en`            missing
#>   <chr>    <chr>  <chr>                   <dbl>
#> 1 gndr     Male   Male                        0
#> 2 gndr     Female Female                      0
#> 3 gndr     -77    Don’t want to answer        1
#> 4 prg_ever 0      never pregnant              0
#> 5 prg_ever 1      pregnant once or more       0
#> 6 prg_ever 2      currently pregnant          0
#> 7 prg_ever 8      Don’t want to answer        1
#> 8 prg_ever 9      Don’t know                  1
#> 9 prg_ever -7     not applicable              1
#>