Transform column(s) of a data dictionary from long format to wide format

Transforms column(s) of a data dictionary from long format to wide format. If a taxonomy is provided, the corresponding columns in the data dictionary will be converted to a format with the taxonomy expanded. This operation is equivalent to performing a tidyr::pivot_wider() on these columns following the taxonomy structure provided. Variable names in the data dictionary must be unique.

data_dict_pivot_wider(data_dict, taxonomy = NULL)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed.

taxonomy

A tibble identifying the scheme used for variables classification.

Value

A list of tibble(s) identifying a data dictionary.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# use DEMO_files provided by the package

taxonomy <- DEMO_files$taxonomy_opal_mlstr
data_dict <- DEMO_files$dd_TOKYO_format_maelstrom_tagged
data_dict_pivot_wider(data_dict, taxonomy)

}
#> $Variables
#> # A tibble: 10 × 13
#>    index name         `label:en`            `description:en`     valueType unit 
#>    <chr> <chr>        <chr>                 <chr>                <chr>     <chr>
#>  1 1     part_id      id of the participant id of the participa… text      NA   
#>  2 2     gndr         gndr                  gender of the parti… text      NA   
#>  3 3     height       height                height of the parti… integer   cm   
#>  4 4     weight_ms    weight_ms             weight of the parti… integer   kg   
#>  5 5     weight_dc    weight_dc             weight of the parti… decimal   kg   
#>  6 6     dob          dob                   date of birth of th… date      years
#>  7 7     collect_date collect_date          date of collection   datetime  NA   
#>  8 8     prg_ever     prg_ever              whether the partici… integer   NA   
#>  9 9     empty        empty                 empty column         integer   NA   
#> 10 10    opentext     opentext              open text            text      NA   
#> # ℹ 7 more variables: `Mlstr_additional::Source` <chr>,
#> #   `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::Administrative_information` <chr>,
#> #   `Mlstr_area::Sociodemographic_economic_characteristics` <chr>,
#> #   `Mlstr_area::Physical_measures` <chr>, `Mlstr_area::Reproduction` <chr>,
#> #   `Mlstr_area::Non_pharmacological_interventions` <chr>
#> 
#> $Categories
#> # A tibble: 11 × 4
#>    variable  name   `label:en`            missing
#>    <chr>     <chr>  <chr>                 <lgl>  
#>  1 gndr      Male   Male                  FALSE  
#>  2 gndr      Female Female                FALSE  
#>  3 gndr      -77    Don’t want to answer  TRUE   
#>  4 weight_ms -88    Don’t want to answer  TRUE   
#>  5 weight_ms -99    Don’t know            TRUE   
#>  6 prg_ever  0      never pregnant        FALSE  
#>  7 prg_ever  1      pregnant once or more FALSE  
#>  8 prg_ever  2      currently pregnant    FALSE  
#>  9 prg_ever  8      Don’t want to answer  TRUE   
#> 10 prg_ever  9      Don’t know            TRUE   
#> 11 prg_ever  -7     not applicable        TRUE   
#>