Generate a web application (bookdown) report of list of a datasets

Generates a visual report for a dataset in an HTML bookdown document, showing descriptive statistics for each variable to facilitate the assessment of input dataset. Statistics and figures are generated according to their valueType. This report can be used to assist the user in the assessment of the dataset structure, fields investigation (mandatory or not), coherence across elements and taxonomy or standard evaluation. The summaries and figures associated provide dataset composition, with observation distribution and descriptive statistics.

dataset_visualize(
  dataset = tibble(id = as.character()),
  data_dict = data_dict_extract(dataset),
  group_by = NULL,
  to,
  taxonomy = NULL,
  .dataset_name = NULL,
  .keep_files = TRUE,
  .summary_var = NULL
)

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

group_by

A character string of one column in the dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

to

A character string identifying the folder path where the bookdown report will be saved.

taxonomy

A data-frame or data-frame extension (e.g. a tibble), identifying the scheme used for variables classification as a tibble.

.dataset_name

A character string specifying the name of the dataset (internally used in the function madshapR::dataset_summarise()).

.keep_files

whether to keep the R-markdown files. TRUE by default.

.summary_var

A summary list which provides summary of the variables ( If provided, the function will not do this action internally, enhancing the speed of completion)

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'index.html'. Or use open_visual_report()

Details

A dossier must be a named list containing at least one data-frame or data-frame extension (e.g. a tibble), each of them being datasets. The name of each tibble will be use as the reference name of the dataset. A data dictionary-like structure must be a list of at least one or two data-frame or data-frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

See also

Examples

{

# use DEMO_files provided by the package
library(dplyr)
library(fabR)

###### Example : Combine functions and summarise datasets.
data_dict <- as_data_dict_mlstr(DEMO_files$dd_TOKYO_format_maelstrom_tagged)
dataset <-
  DEMO_files$dataset_TOKYO %>%
  valueType_adjust(from = data_dict) %>%
  data_dict_apply(data_dict)

tempdir <- tempdir()

dataset_visualize(dataset,group_by = gndr,to = tempdir)
  
open_visual_report(tempdir)

}
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>     The data dictionary contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for numerical variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#>   |                                                           |                                                   |   0%  |                                                           |.                                                  |   1%                     |                                                           |.                                                  |   2% (unnamed-chunk-1)   |                                                           |..                                                 |   4%                     |                                                           |..                                                 |   5% (unnamed-chunk-2)   |                                                           |...                                                |   6%                     |                                                           |....                                               |   7% (unnamed-chunk-3)   |                                                           |....                                               |   8%                     |                                                           |.....                                              |   9% (unnamed-chunk-4)   |                                                           |.....                                              |  11%                     |                                                           |......                                             |  12% (unnamed-chunk-5)   |                                                           |.......                                            |  13%                     |                                                           |.......                                            |  14% (unnamed-chunk-6)   |                                                           |........                                           |  15%                     |                                                           |........                                           |  16% (figures-plot12-1)  |                                                           |.........                                          |  18%                     |                                                           |..........                                         |  19% (unnamed-chunk-7)   |                                                           |..........                                         |  20%                     |                                                           |...........                                        |  21% (unnamed-chunk-8)   |                                                           |...........                                        |  22%                     |                                                           |............                                       |  24% (unnamed-chunk-9)   |                                                           |.............                                      |  25%                     |                                                           |.............                                      |  26% (unnamed-chunk-10)  |                                                           |..............                                     |  27%                     |                                                           |..............                                     |  28% (figures-plot12-2)  |                                                           |...............                                    |  29%                     |                                                           |................                                   |  31% (unnamed-chunk-11)  |                                                           |................                                   |  32%                     |                                                           |.................                                  |  33% (unnamed-chunk-12)  |                                                           |.................                                  |  34%                     |                                                           |..................                                 |  35% (unnamed-chunk-13)  |                                                           |...................                                |  36%                     |                                                           |...................                                |  38% (figures-plot12-3)  |                                                           |....................                               |  39%                     |                                                           |....................                               |  40% (unnamed-chunk-14)  |                                                           |.....................                              |  41%                     |                                                           |......................                             |  42% (unnamed-chunk-15)  |                                                           |......................                             |  44%                     |                                                           |.......................                            |  45% (unnamed-chunk-16)  |                                                           |.......................                            |  46%                     |                                                           |........................                           |  47% (unnamed-chunk-17)  |                                                           |.........................                          |  48%                     |                                                           |.........................                          |  49% (figures-plot12-4)  |                                                           |..........................                         |  51%                     |                                                           |..........................                         |  52% (unnamed-chunk-18)  |                                                           |...........................                        |  53%                     |                                                           |............................                       |  54% (unnamed-chunk-19)  |                                                           |............................                       |  55%                     |                                                           |.............................                      |  56% (unnamed-chunk-20)  |                                                           |.............................                      |  58%                     |                                                           |..............................                     |  59% (figures-plot12-5)  |                                                           |...............................                    |  60%                     |                                                           |...............................                    |  61% (unnamed-chunk-21)  |                                                           |................................                   |  62%                     |                                                           |................................                   |  64% (unnamed-chunk-22)  |                                                           |.................................                  |  65%                     |                                                           |..................................                 |  66% (unnamed-chunk-23)  |                                                           |..................................                 |  67%                     |                                                           |...................................                |  68% (figures-plot12-6)  |                                                           |...................................                |  69%                     |                                                           |....................................               |  71% (unnamed-chunk-24)  |                                                           |.....................................              |  72%                     |                                                           |.....................................              |  73% (unnamed-chunk-25)  |                                                           |......................................             |  74%                     |                                                           |......................................             |  75% (unnamed-chunk-26)  |                                                           |.......................................            |  76%                     |                                                           |........................................           |  78% (unnamed-chunk-27)  |                                                           |........................................           |  79%                     |                                                           |.........................................          |  80% (figures-plot12-7)  |                                                           |.........................................          |  81%                     |                                                           |..........................................         |  82% (unnamed-chunk-28)  |                                                           |...........................................        |  84%                     |                                                           |...........................................        |  85% (unnamed-chunk-29)  |                                                           |............................................       |  86%                     |                                                           |............................................       |  87% (unnamed-chunk-30)  |                                                           |.............................................      |  88%                     |                                                           |..............................................     |  89% (figures-plot12-8)  |                                                           |..............................................     |  91%                     |                                                           |...............................................    |  92% (unnamed-chunk-31)  |                                                           |...............................................    |  93%                     |                                                           |................................................   |  94% (unnamed-chunk-32)  |                                                           |.................................................  |  95%                     |                                                           |.................................................  |  96% (unnamed-chunk-33)  |                                                           |.................................................. |  98%                     |                                                           |.................................................. |  99% (figures-plot12-9)  |                                                           |...................................................| 100%                                                                                                                                
#> "C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/pandoc" +RTS -K512m -RTS bookdownproj.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output bookdownproj.html --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\rmarkdown\lua\custom-environment.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\latex-div.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\anchor-sections.lua" --metadata-file "C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c\filefa8bd6433c" --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\templates\gitbook.html" --highlight-style pygments --number-sections --css style.css --mathjax --include-in-header "C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c\rmarkdown-strfa836f505d.html" 
#> 
#> 
#> To edit your file, You can use the function `open_visual_report('C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c')`
#> (Compatibility tested on Chrome and Mozilla)
#>