Remove labels (attributes) from a data frame, leaving its unlabelled columns

Removes any attributes attached to a data frame. Any value in columns will be preserved. Any 'Date' (typeof) column will be recast as character to preserve information.

dataset_zap_data_dict(dataset)

Arguments

dataset

A data frame identifying the input dataset observations associated to its data dictionary.

Value

A data frame identifying a dataset.

Details

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

See also

Examples

{

# use madshapR_DEMO provided by the package

dataset <- madshapR_DEMO$dataset_TOKYO
data_dict <- as_data_dict_mlstr(madshapR_DEMO$data_dict_TOKYO)
dataset <- data_dict_apply(dataset,data_dict)
dataset_zap_data_dict(dataset)

}
#> # A tibble: 50 × 9
#>    part_id gndr   height weight_ms weight_dc dob        prg_ever empty opentext 
#>  * <chr>   <chr>   <dbl>     <dbl>     <dbl> <chr>         <dbl> <lgl> <chr>    
#>  1 ID001   Male      191        63        NA 3/22/1990        -7 NA    All chil…
#>  2 ID002   Female    176        NA        65 8/15/2001         0 NA    grow up,…
#>  3 ID003   Female    154        NA       121 12/17/1996        2 NA    years ol…
#>  4 ID004   Female    167       -88        NA 6/13/1990         1 NA    flower a…
#>  5 ID005   Female    185        NA        45 12/17/1996        8 NA    rather d…
#>  6 ID006   -77       171        57        NA 3/31/1981        -7 NA    cried, O…
#>  7 ID007   Female    185        NA        58 4/19/1988         9 NA    that pas…
#>  8 ID008   Female    171        NA        59 NA                2 NA    that she…
#>  9 ID009   -77       169        52        NA 3/14/1976        -7 NA    beginnin…
#> 10 ID010   Male      179        NA        62 10/19/1993       -7 NA    All chil…
#> # ℹ 40 more rows