Generate a report and summary of a dossier (list of datasets)

Assesses and summarizes the content and structure of a dossier (list of datasets) and reports potential issues to facilitate the assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. The summary provides additional information about variable distributions and descriptive statistics. This report is compatible with Excel and can be exported as an Excel spreadsheet.

dossier_summarize(
  dossier,
  group_by = NULL,
  taxonomy = NULL,
  valueType_guess = FALSE
)

Arguments

dossier

List of tibble, each of them being datasets.

group_by

A character string of one column in the dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

taxonomy

A tibble identifying the scheme used for variables classification.

valueType_guess

Whether the output should include a more accurate valueType that could be applied to the dataset. TRUE by default.

Value

A list of tibbles of report for each listed dataset.

Details

A dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being datasets. The name of each tibble will be use as the reference name of the dataset. This report is compatible with Excel and can be exported as an Excel spreadsheet.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondence with R data types are available using valueType_list.

Examples

{

# use DEMO_files provided by the package
library(dplyr)

###### Example 1: Combine functions and summarise datasets.
dossier <- list(iris = tibble())

dossier_summary <- dossier_summarize(dossier)
glimpse(dossier_summary)

}
#> - DOSSIER SUMMARY: -----------------------------------------------------
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>     The data dictionary contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: iris (empty dataset) --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>     The dataset contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: iris (empty dataset) --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise global information (Overview)
#>     Generate report
#> List of 1
#>  $ iris:List of 1
#>   ..$ Overview: tibble [15 × 2] (S3: tbl_df/tbl/data.frame)