Built-in material allowing the user to test the package with demo data

Built-in tibbles and lists allowing the user to test the package with demo material.

madshapR_DEMO

Format

list

A list with 20 elements used for testing the package (data frames and lists):

data_dict_MELBOURNE

Data dictionary of Melbourne dataset

data_dict_PARIS

Data dictionary of Paris dataset

data_dict_PARIS - flatten

Data dictionary of Paris in preprocessed format

data_dict_TOKYO

Data dictionary of Tokyo dataset

data_dict_TOKYO - with errors

Tagged data dictionary of Tokyo dataset containing errors for testing purpose

data_dictt_TOKYO - with errors in data

data dictionary of Tokyo dataset containing errors with for testing purpose

dataset_MELBOURNE

Dataset of Melbourne

dataset_PARIS

Dataset of Paris

dataset_TOKYO

Dataset of Tokyo

data_dict_TOKYO - with errors in data

Dataset of Tokyo containing errors for testing purpose

...

Examples

{

 print(madshapR_DEMO$dataset_TOKYO)

}
#> # A tibble: 50 × 9
#>    part_id gndr   height weight_ms weight_dc dob        prg_ever empty opentext 
#>    <chr>   <chr>   <dbl>     <dbl>     <dbl> <chr>         <dbl> <lgl> <chr>    
#>  1 ID001   Male      191        63        NA 3/22/1990        -7 NA    All chil…
#>  2 ID002   Female    176        NA        65 8/15/2001         0 NA    grow up,…
#>  3 ID003   Female    154        NA       121 12/17/1996        2 NA    years ol…
#>  4 ID004   Female    167       -88        NA 6/13/1990         1 NA    flower a…
#>  5 ID005   Female    185        NA        45 12/17/1996        8 NA    rather d…
#>  6 ID006   -77       171        57        NA 3/31/1981        -7 NA    cried, O…
#>  7 ID007   Female    185        NA        58 4/19/1988         9 NA    that pas…
#>  8 ID008   Female    171        NA        59 NA                2 NA    that she…
#>  9 ID009   -77       169        52        NA 3/14/1976        -7 NA    beginnin…
#> 10 ID010   Male      179        NA        62 10/19/1993       -7 NA    All chil…
#> # ℹ 40 more rows