Built-in material allowing the user to test the package with example data

Example datasets and data dictionaries, and taxonomy, to provide illustrative examples of objects used by madshapR.

madshapR_examples

Format

list

A list with 9 elements (data frames and lists) providing example objects for testing the package:

dataset_example

Dataset for example dataset

data_dictionary_example - as_data_dict

Data dictionary for example dataset where the structure is a data dictionary

data_dictionary_example - as_data_dict_mlstr

Data dictionary for example dataset where the structure is a data dictionary compliant with Maelstrom

dataset_example - errors with data

Dataset of example with errors with example data dictionary

data_dictionary_example - errors with data

Data Dictionary for example dataset with errors with example dataset

data_dictionary_example - errors

Data dictionary for example dataset with errors

data_dictionary_example - collapsed

Data dictionary for example with collapsed categories

taxonomy_example

Taxonomy for example dataset

summary - dataset_example

Dataset example summary

...

Examples

{
 
library(dplyr)
 
head(madshapR_examples$`dataset_example`)
glimpse(madshapR_examples$`data_dictionary_example`)

}
#> List of 2
#>  $ Variables : tibble [9 × 8] (S3: tbl_df/tbl/data.frame)
#>   ..$ index                : num [1:9] 1 2 3 4 5 6 7 8 9
#>   ..$ name                 : chr [1:9] "part_id" "gndr" "height" "weight_ms" ...
#>   ..$ label:en             : chr [1:9] "id of the participant" "gndr" "height" "weight_ms" ...
#>   ..$ description:en       : chr [1:9] "id of the participant" "gender of the participant" "height of the participant" "weight of the participant - measured" ...
#>   ..$ valueType            : chr [1:9] "text" "integer" "integer" "integer" ...
#>   ..$ unit                 : chr [1:9] NA NA "cm" "kg" ...
#>   ..$ datacollection::type : chr [1:9] "automatic" "declared" "declared" "measured" ...
#>   ..$ datacollection::level: chr [1:9] "high" "high" "moderate" "moderate" ...
#>  $ Categories: tibble [11 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ variable: chr [1:11] "gndr" "gndr" "gndr" "weight_ms" ...
#>   ..$ name    : chr [1:11] "1" "2" "-77" "-88" ...
#>   ..$ label:en: chr [1:11] "Male" "Female" "Don’t want to answer" "Don’t want to answer" ...
#>   ..$ missing : logi [1:11] FALSE FALSE TRUE TRUE TRUE FALSE ...