Download tables from an Opal project as a dossier

Downloads a dossier or dataset(s) from Opal project to local environment. It is a wrapper of opalr::opal.table_get() and opalr::opal.table_dictionary_get().

opal_tables_pull(
  opal,
  project,
  table_list = NULL,
  content = c("dataset", "data_dict"),
  keep_as_dossier = TRUE,
  .remove_id = FALSE
)

Arguments

opal

Opal login attributes.

project

A character string specifying the Opal project name.

table_list

A vector character string specifying Opal tables name.

content

A vector of character string which indicates if the function returns a dataset, or data dictionary. Default is 'dataset'.

keep_as_dossier

whether to return a dossier or a dataset if there is only one table. TRUE by default, if FALSE returns dataset.

.remove_id

whether to return the id column created in Opal or not. TRUE by default.

Value

R objects (tibbles and list of tibbles) representing tables and their respective data dictionary.

Details

The user must be allowed to interact with their Opal. The errors provided may be associated to the handler or the read/write permissions to Opal. The user must have adequate credentials to interact with their Opal environment.

See also

Examples

if (FALSE) {

#' # use DEMO_files provided by the package
library(opalr)
library(stringr)

dossier <- 
  DEMO_files[stringr::str_detect(names(DEMO_files),"dataset_MELBOURNE")]

opal <- 
  opal.login('administrator','password', 
    url ='https://opal-demo.maelstrom-research.org')
  
tempdir <- basename(tempdir())
try(opal_project_create(opal, tempdir))

try(
  opal_tables_push(
  opal, dossier,project_name = tempdir, .force = TRUE, .overwrite = TRUE))
  
###### Example 1: pull a table from a project.
try(
  opal_tables_pull(
  opal,project = tempdir,table_list = 'dataset_MELBOURNE_1'))

###### Example 2: pull a dossier from a project.
try(
  opal_tables_pull(
  opal, project = tempdir))

###### Example 3: pull a data dictionary from a project.
try(
  opal_tables_pull(
  opal, project = tempdir, content = 'data_dict'))

}