Open a visual report in a browser

The visual report previously generated in an HTML bookdown document can be opened using this short-cut function. This report can be used to assist the user in the assessment of the dataset structure, fields investigation (mandatory or not), coherence across elements and taxonomy or standard evaluation. The summaries and figures associated provide dataset composition, with observation distribution and descriptive statistics.

open_visual_report(report_name)

Arguments

report_name

A character string specifying the name of the report (a folder in users environment) to be opened.

See also

Examples

{

# use DEMO_files provided by the package
library(dplyr)

###### Example 1: any data-frame (or tibble) can be a dataset by definition.
tempdir <- tempdir()
dataset_visualize(iris, to = tempdir)
open_visual_report(tempdir)

}
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: iris --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: iris --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for numerical variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#>   |                                                           |                                                   |   0%  |                                                           |.                                                  |   2%                     |                                                           |..                                                 |   4% (unnamed-chunk-1)   |                                                           |...                                                |   6%                     |                                                           |....                                               |   8% (unnamed-chunk-2)   |                                                           |.....                                              |  10%                     |                                                           |......                                             |  12% (unnamed-chunk-3)   |                                                           |.......                                            |  14%                     |                                                           |........                                           |  16% (unnamed-chunk-4)   |                                                           |.........                                          |  18%                     |                                                           |..........                                         |  20% (unnamed-chunk-5)   |                                                           |...........                                        |  22%                     |                                                           |............                                       |  24% (unnamed-chunk-6)   |                                                           |..............                                     |  27%                     |                                                           |...............                                    |  29% (figures-plot12-1)  |                                                           |................                                   |  31%                     |                                                           |.................                                  |  33% (unnamed-chunk-7)   |                                                           |..................                                 |  35%                     |                                                           |...................                                |  37% (unnamed-chunk-8)   |                                                           |....................                               |  39%                     |                                                           |.....................                              |  41% (unnamed-chunk-9)   |                                                           |......................                             |  43%                     |                                                           |.......................                            |  45% (figures-plot12-2)  |                                                           |........................                           |  47%                     |                                                           |.........................                          |  49% (unnamed-chunk-10)  |                                                           |..........................                         |  51%                     |                                                           |...........................                        |  53% (unnamed-chunk-11)  |                                                           |............................                       |  55%                     |                                                           |.............................                      |  57% (unnamed-chunk-12)  |                                                           |..............................                     |  59%                     |                                                           |...............................                    |  61% (figures-plot12-3)  |                                                           |................................                   |  63%                     |                                                           |.................................                  |  65% (unnamed-chunk-13)  |                                                           |..................................                 |  67%                     |                                                           |...................................                |  69% (unnamed-chunk-14)  |                                                           |....................................               |  71%                     |                                                           |.....................................              |  73% (unnamed-chunk-15)  |                                                           |.......................................            |  76%                     |                                                           |........................................           |  78% (figures-plot12-4)  |                                                           |.........................................          |  80%                     |                                                           |..........................................         |  82% (unnamed-chunk-16)  |                                                           |...........................................        |  84%                     |                                                           |............................................       |  86% (unnamed-chunk-17)  |                                                           |.............................................      |  88%                     |                                                           |..............................................     |  90% (unnamed-chunk-18)  |                                                           |...............................................    |  92%                     |                                                           |................................................   |  94% (unnamed-chunk-19)  |                                                           |.................................................  |  96%                     |                                                           |.................................................. |  98% (figures-plot12-5)  |                                                           |...................................................| 100%                                                                                                                                
#> "C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/pandoc" +RTS -K512m -RTS bookdownproj.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output bookdownproj.html --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\rmarkdown\lua\custom-environment.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\latex-div.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\anchor-sections.lua" --metadata-file "C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c\filefa846fb1cd4" --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\templates\gitbook.html" --highlight-style pygments --number-sections --css style.css --mathjax --include-in-header "C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c\rmarkdown-strfa868972cf5.html" 
#> 
#> 
#> To edit your file, You can use the function `open_visual_report('C:\Users\guill\AppData\Local\Temp\Rtmp4Sux3c')`
#> (Compatibility tested on Chrome and Mozilla)
#>