Provide descriptive statistics for variables in a dataset

Summarises (in a tibble) the columns in a dataset and its data dictionary (if any). The summary provides information about quality, type, composition, and descriptive statistics of variables. Statistics are generated by valueType.

summary_variables(dataset = NULL, data_dict = NULL, .dataset_preprocess = NULL)

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

.dataset_preprocess

A tibble which provides summary of the variables (used for internal processes and programming).

Value

A tibble providing statistical description of variables present in a dataset.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

Examples

{

###### Example : any data frame (or tibble) can be a dataset by definition.
.dataset_preprocess <- dataset_preprocess(iris)
summary_variables(.dataset_preprocess = .dataset_preprocess)

}
#> # A tibble: 5 × 9
#>   name      Quality assessment c…¹ `Categorical variable` Total number of obse…²
#>   <chr>     <chr>                  <chr>                                   <dbl>
#> 1 Sepal.Le… NA                     no                                        150
#> 2 Sepal.Wi… NA                     no                                        150
#> 3 Petal.Le… NA                     no                                        150
#> 4 Petal.Wi… NA                     no                                        150
#> 5 Species   NA                     yes                                       150
#> # ℹ abbreviated names: ¹​`Quality assessment comment`,
#> #   ²​`Total number of observations`
#> # ℹ 5 more variables: `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>