Provide statistical description of 'date' variables of a dataset

Generates a tibble that summarises all 'date' columns in a dataset with (if any) its data dictionary. The data dictionary (if present separates observations between open values, missing values, categorical values, and categorical missing values (which corresponds to the 'missing' column in the Categories' sheet). Statistics are generated according to their valueType which can be 'date'. This summary can be used to assist the user in the assessment of the dataset composition, with observation distribution and descriptive statistics.

summary_variables_date(dataset = NULL, data_dict = NULL, .resume_var = NULL)

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

.resume_var

A tibble which provides summary of the variables (for development purpose only)

Value

A tibble providing statistical description of 'date' variables present in a dataset.

Details

A data dictionary-like structure must be a list of at least one or two data-frame or data-frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

Examples

{
   
###### Example : any data-frame (or tibble) can be a dataset by definition.
library(dplyr)
library(fabR)

.resume_var <- 
  storms %>%
    sample_n(50) %>%
    mutate(date_storm = as_any_date(paste(year, month, day,"-"),"ymd")) %>%
    select(date_storm) %>%
    resume_variables

summary_variables_date(.resume_var = .resume_var)

}
#> # A tibble: 1 × 10
#>   name  `Lowest date` `Highest date` `MIN\n(year)` `Q1\n(year)` `MEDIAN\n(year)`
#>   <chr> <date>        <date>                 <dbl>        <dbl>            <dbl>
#> 1 date… 1975-10-02    2021-09-30              1975         1990             2002
#> # ℹ 4 more variables: `Q3\n(year)` <dbl>, `MAX\n(year)` <dbl>,
#> #   `MEAN\n(year)` <dbl>, `Span\n(year)` <dbl>