Provide descriptive statistics for variables of type 'datetime' in a dataset

Summarizes (in a data frame) the columns of type 'datetime' in a dataset and its data dictionary (if any). The summary provides information about quality, type, composition, and descriptive statistics of variables. Statistics are generated by valueType.

summary_variables_datetime(
  dataset = NULL,
  data_dict = NULL,
  dataset_preprocess = .dataset_preprocess,
  .dataset_preprocess = NULL
)

Arguments

dataset

A data frame identifying the input dataset observations associated to its data dictionary.

data_dict

A list of data frame(s) representing meta data of an associated dataset. Automatically generated if not provided.

dataset_preprocess

A data frame which provides summary of the variables (for internal processes and programming).

.dataset_preprocess

Value

A data frame providing statistical description of 'datetime' variables present in a dataset.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

Examples

{
   
###### Example : any data frame can be a dataset by definition.
library(dplyr)
library(fabR)
library(lubridate)

dataset_preprocess <- 
 storms %>%
   sample_n(50) %>%
   mutate(date_storm = as_datetime(paste(year, month, day,"-"))) %>%
   select(date_storm) %>%
   dataset_preprocess

summary_variables_datetime(dataset_preprocess = dataset_preprocess)

}
#> 
#> Attaching package: 'lubridate'
#> The following objects are masked from 'package:base':
#> 
#>     date, intersect, setdiff, union
#> # A tibble: 1 × 3
#>   name       `Most seen value(s)`                              `Rarest value(s)`
#>   <chr>      <chr>                                             <chr>            
#> 1 date_storm 1975-09-24 ; 1981-09-09 ; 1982-06-02 ; 1982-09-1… 1975-09-24 ; 198…