Attribute the valueType from a data dictionary to a dataset, or vice versa

Takes the valueType of the input (from) and attributes it to the output (to). The parameters 'from' and 'to' can be either a dataset or a data dictionary. Depending on the input provided, the valueType replaced is either in the 'valueType' column of a data dictionary or cast to a column in a dataset. If 'to' is not provided, the function calls valueType_self_adjust() instead. The possible values returned are 'date', 'boolean', 'integer', 'decimal', and text'.

valueType_adjust(from, to = NULL)

Arguments

from

Object to be adjusted. Can be either a dataset or a data dictionary.

to

Object to be adjusted. Can be either a dataset or a data dictionary. NULL by default.

Value

Either a tibble, identifying the dataset, or a list of tibble(s) identifying a data dictionary, depending which is 'to'.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', 'datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondance with R data types are available using valueType_list.

See also

Examples

{

# use DEMO_files provided by the package

data_dict <- DEMO_files$dd_TOKYO_format_maelstrom_tagged
dataset <- DEMO_files$dataset_TOKYO

###### Example 1: The valueType of a dataset can be adjusted using 
# the valueType present in the data dictionary.
valueType_adjust(from = data_dict,to = dataset)

###### Example 2: The valueType of a dataset can be adjusted using 
# the valueType present in the data dictionary. Here the valueType is 'text'
data_dict[['Variables']]$valueType <- 'text'
valueType_adjust(from = data_dict,to = dataset)

###### Example 1: The valueType of a data dictionary can be adjusted using 
# the valueType of each column of the dataset.
data_dict[['Variables']]$valueType <- 'text'
dataset <- valueType_self_adjust(dataset)
valueType_adjust(from = dataset,to = data_dict)
}
#> $Variables
#> # A tibble: 9 × 16
#>   index name  `label:en` `description:en` valueType unit  Mlstr_additional::So…¹
#>   <dbl> <chr> <chr>      <chr>            <chr>     <chr> <chr>                 
#> 1     1 part… id of the… id of the parti… text      NA    Questionnaire         
#> 2     2 gndr  gndr       gender of the p… text      NA    Questionnaire         
#> 3     3 heig… height     height of the p… integer   cm    Physical_measure      
#> 4     4 weig… weight_ms  weight of the p… integer   kg    Physical_measure      
#> 5     5 weig… weight_dc  weight of the p… integer   kg    Physical_measure      
#> 6     6 dob   dob        date of birth o… date      years Questionnaire         
#> 7     7 prg_… prg_ever   whether the par… integer   NA    Questionnaire         
#> 8     8 empty empty      empty column     boolean   NA    Questionnaire         
#> 9     9 open… opentext   open text        text      NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 9 more variables: `Mlstr_additional::Target` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <lgl>,
#> #   `Mlstr_area::2` <chr>, `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>, typeof <chr>
#> 
#> $Categories
#> # A tibble: 11 × 4
#>    variable  name   `label:en`            missing
#>    <chr>     <chr>  <chr>                 <lgl>  
#>  1 gndr      Male   Male                  FALSE  
#>  2 gndr      Female Female                FALSE  
#>  3 gndr      -77    Don’t want to answer  TRUE   
#>  4 weight_ms -88    Don’t want to answer  TRUE   
#>  5 weight_ms -99    Don’t know            TRUE   
#>  6 prg_ever  0      never pregnant        FALSE  
#>  7 prg_ever  1      pregnant once or more FALSE  
#>  8 prg_ever  2      currently pregnant    FALSE  
#>  9 prg_ever  8      Don’t want to answer  TRUE   
#> 10 prg_ever  9      Don’t know            TRUE   
#> 11 prg_ever  -7     not applicable        TRUE   
#>