Return the valueType of an object

Determines the valueType of an object based on typeof() and class(). The possible values returned are 'date', 'boolean', 'integer', 'decimal', and 'text'.

valueType_of(x)

Arguments

x

Object. Can be a vector.

Value

A character string which is the valueType of the input object.

Details

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondence with R data types are available using valueType_list.

See also

Examples

{

# use DEMO_files provided by the package

dataset <- DEMO_files$dataset_MELBOURNE_1
valueType_of(dataset$Gender)
valueType_of(iris$Sepal.Length)

}
#> [1] "decimal"