Built-in material allowing the user to test the package with demo data

Built-in tibbles and lists allowing the user to test the package with demo material.

DEMO_files

Format

list

A list with 22 elements used for testing the package (data frames and lists):

DEMO_data_processing_elements - final

Data processing element without error

DEMO_data_processing_elements - with error

Data processing element containing errors

DEMO_data_processing_elements - work in progress

Data processing element in construction

DEMO_dataschema

DataSchema used, combined with data processing elements and study-specific material

dd_MELBOURNE_1_format_maelstrom

Data dictionary (1) of Melbourne dataset

dd_MELBOURNE_2_format_maelstrom

Data dictionary (2) of Melbourne dataset

dd_PARIS_format_maelstrom

Data dictionary of Paris dataset

dd_PARIS_format_preprocessed - ERROR

Data dictionary of Paris dataset containing errors

dd_PARIS_format_preprocessed

Data dictionary of Paris in preprocessed format

dd_TOKYO_format_maelstrom_tagged - ERROR WITH DATA

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_maelstrom_tagged - ERROR

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_maelstrom_tagged

Data dictionary of Tokyo dataset

dd_TOKYO_format_opal_tagged - ERROR WITH TAXO

Data dictionary of Tokyo dataset containing errors

dd_TOKYO_format_opal_tagged

Data dictionary of Tokyo dataset opal format

dataset_MELBOURNE_1

Dataset of Melbourne (1)

dataset_MELBOURNE_2

Dataset of Melbourne (2)

dataset_PARIS

Dataset of Paris

dataset_TOKYO - ERROR WITH DATA

Dataset of Tokyo containing errors

dataset_TOKYO

Dataset of Tokyo

PARIS_taxonomy

Taxonomy specific to Paris dataset

taxonomy_mlstr

Maelstrom Taxonomy

taxonomy_opal

Opal Taxonomy

...

Examples

{

 print(DEMO_files$dataset_TOKYO)

}
#> # A tibble: 50 × 9
#>    part_id gndr   height weight_ms weight_dc dob        prg_ever empty opentext 
#>    <chr>   <chr>   <dbl>     <dbl>     <dbl> <chr>         <dbl> <lgl> <chr>    
#>  1 ID001   Male      191        63        NA 3/22/1990        -7 NA    All chil…
#>  2 ID002   Female    176        NA        65 8/15/2001         0 NA    grow up,…
#>  3 ID003   Female    154        NA       121 12/17/1996        2 NA    years ol…
#>  4 ID004   Female    167       -88        NA 6/13/1990         1 NA    flower a…
#>  5 ID005   Female    185        NA        45 12/17/1996        8 NA    rather d…
#>  6 ID006   -77       171        57        NA 3/31/1981        -7 NA    cried, O…
#>  7 ID007   Female    185        NA        58 4/19/1988         9 NA    that pas…
#>  8 ID008   Female    171        NA        59 NA                2 NA    that she…
#>  9 ID009   -77       169        52        NA 3/14/1976        -7 NA    beginnin…
#> 10 ID010   Male      179        NA        62 10/19/1993       -7 NA    All chil…
#> # ℹ 40 more rows