Validate and coerce an object to dataset format

Confirms that the input object is a valid dataset and returns it as a dataset with the appropriate madshapR::class attribute. This function mainly helps validate inputs within other functions of the package but could be used to check if a dataset is valid.

as_dataset(object, col_id = NULL)

Arguments

object

A potential dataset to be coerced.

col_id

A character string specifying the name(s) of the column(s) which refer to key identifier of the dataset. The column(s) can be named or indicated by position.

Value

A list of data frame(s) with Rmonize::class 'dataset'.

Details

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

Examples

{

# use madshapR_DEMO provided by the package

###### Example 1: a dataset can have an id column(s) which is specified as
# an attribute. 
dataset <- as_dataset(madshapR_DEMO$dataset_MELBOURNE, col_id = "id")

###### Example 2: any data frame can be a dataset by definition.
as_dataset(iris, col_id = "Species")
}
#>        Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1       setosa          5.1         3.5          1.4         0.2
#> 2       setosa          4.9         3.0          1.4         0.2
#> 3       setosa          4.7         3.2          1.3         0.2
#> 4       setosa          4.6         3.1          1.5         0.2
#> 5       setosa          5.0         3.6          1.4         0.2
#> 6       setosa          5.4         3.9          1.7         0.4
#> 7       setosa          4.6         3.4          1.4         0.3
#> 8       setosa          5.0         3.4          1.5         0.2
#> 9       setosa          4.4         2.9          1.4         0.2
#> 10      setosa          4.9         3.1          1.5         0.1
#> 11      setosa          5.4         3.7          1.5         0.2
#> 12      setosa          4.8         3.4          1.6         0.2
#> 13      setosa          4.8         3.0          1.4         0.1
#> 14      setosa          4.3         3.0          1.1         0.1
#> 15      setosa          5.8         4.0          1.2         0.2
#> 16      setosa          5.7         4.4          1.5         0.4
#> 17      setosa          5.4         3.9          1.3         0.4
#> 18      setosa          5.1         3.5          1.4         0.3
#> 19      setosa          5.7         3.8          1.7         0.3
#> 20      setosa          5.1         3.8          1.5         0.3
#> 21      setosa          5.4         3.4          1.7         0.2
#> 22      setosa          5.1         3.7          1.5         0.4
#> 23      setosa          4.6         3.6          1.0         0.2
#> 24      setosa          5.1         3.3          1.7         0.5
#> 25      setosa          4.8         3.4          1.9         0.2
#> 26      setosa          5.0         3.0          1.6         0.2
#> 27      setosa          5.0         3.4          1.6         0.4
#> 28      setosa          5.2         3.5          1.5         0.2
#> 29      setosa          5.2         3.4          1.4         0.2
#> 30      setosa          4.7         3.2          1.6         0.2
#> 31      setosa          4.8         3.1          1.6         0.2
#> 32      setosa          5.4         3.4          1.5         0.4
#> 33      setosa          5.2         4.1          1.5         0.1
#> 34      setosa          5.5         4.2          1.4         0.2
#> 35      setosa          4.9         3.1          1.5         0.2
#> 36      setosa          5.0         3.2          1.2         0.2
#> 37      setosa          5.5         3.5          1.3         0.2
#> 38      setosa          4.9         3.6          1.4         0.1
#> 39      setosa          4.4         3.0          1.3         0.2
#> 40      setosa          5.1         3.4          1.5         0.2
#> 41      setosa          5.0         3.5          1.3         0.3
#> 42      setosa          4.5         2.3          1.3         0.3
#> 43      setosa          4.4         3.2          1.3         0.2
#> 44      setosa          5.0         3.5          1.6         0.6
#> 45      setosa          5.1         3.8          1.9         0.4
#> 46      setosa          4.8         3.0          1.4         0.3
#> 47      setosa          5.1         3.8          1.6         0.2
#> 48      setosa          4.6         3.2          1.4         0.2
#> 49      setosa          5.3         3.7          1.5         0.2
#> 50      setosa          5.0         3.3          1.4         0.2
#> 51  versicolor          7.0         3.2          4.7         1.4
#> 52  versicolor          6.4         3.2          4.5         1.5
#> 53  versicolor          6.9         3.1          4.9         1.5
#> 54  versicolor          5.5         2.3          4.0         1.3
#> 55  versicolor          6.5         2.8          4.6         1.5
#> 56  versicolor          5.7         2.8          4.5         1.3
#> 57  versicolor          6.3         3.3          4.7         1.6
#> 58  versicolor          4.9         2.4          3.3         1.0
#> 59  versicolor          6.6         2.9          4.6         1.3
#> 60  versicolor          5.2         2.7          3.9         1.4
#> 61  versicolor          5.0         2.0          3.5         1.0
#> 62  versicolor          5.9         3.0          4.2         1.5
#> 63  versicolor          6.0         2.2          4.0         1.0
#> 64  versicolor          6.1         2.9          4.7         1.4
#> 65  versicolor          5.6         2.9          3.6         1.3
#> 66  versicolor          6.7         3.1          4.4         1.4
#> 67  versicolor          5.6         3.0          4.5         1.5
#> 68  versicolor          5.8         2.7          4.1         1.0
#> 69  versicolor          6.2         2.2          4.5         1.5
#> 70  versicolor          5.6         2.5          3.9         1.1
#> 71  versicolor          5.9         3.2          4.8         1.8
#> 72  versicolor          6.1         2.8          4.0         1.3
#> 73  versicolor          6.3         2.5          4.9         1.5
#> 74  versicolor          6.1         2.8          4.7         1.2
#> 75  versicolor          6.4         2.9          4.3         1.3
#> 76  versicolor          6.6         3.0          4.4         1.4
#> 77  versicolor          6.8         2.8          4.8         1.4
#> 78  versicolor          6.7         3.0          5.0         1.7
#> 79  versicolor          6.0         2.9          4.5         1.5
#> 80  versicolor          5.7         2.6          3.5         1.0
#> 81  versicolor          5.5         2.4          3.8         1.1
#> 82  versicolor          5.5         2.4          3.7         1.0
#> 83  versicolor          5.8         2.7          3.9         1.2
#> 84  versicolor          6.0         2.7          5.1         1.6
#> 85  versicolor          5.4         3.0          4.5         1.5
#> 86  versicolor          6.0         3.4          4.5         1.6
#> 87  versicolor          6.7         3.1          4.7         1.5
#> 88  versicolor          6.3         2.3          4.4         1.3
#> 89  versicolor          5.6         3.0          4.1         1.3
#> 90  versicolor          5.5         2.5          4.0         1.3
#> 91  versicolor          5.5         2.6          4.4         1.2
#> 92  versicolor          6.1         3.0          4.6         1.4
#> 93  versicolor          5.8         2.6          4.0         1.2
#> 94  versicolor          5.0         2.3          3.3         1.0
#> 95  versicolor          5.6         2.7          4.2         1.3
#> 96  versicolor          5.7         3.0          4.2         1.2
#> 97  versicolor          5.7         2.9          4.2         1.3
#> 98  versicolor          6.2         2.9          4.3         1.3
#> 99  versicolor          5.1         2.5          3.0         1.1
#> 100 versicolor          5.7         2.8          4.1         1.3
#> 101  virginica          6.3         3.3          6.0         2.5
#> 102  virginica          5.8         2.7          5.1         1.9
#> 103  virginica          7.1         3.0          5.9         2.1
#> 104  virginica          6.3         2.9          5.6         1.8
#> 105  virginica          6.5         3.0          5.8         2.2
#> 106  virginica          7.6         3.0          6.6         2.1
#> 107  virginica          4.9         2.5          4.5         1.7
#> 108  virginica          7.3         2.9          6.3         1.8
#> 109  virginica          6.7         2.5          5.8         1.8
#> 110  virginica          7.2         3.6          6.1         2.5
#> 111  virginica          6.5         3.2          5.1         2.0
#> 112  virginica          6.4         2.7          5.3         1.9
#> 113  virginica          6.8         3.0          5.5         2.1
#> 114  virginica          5.7         2.5          5.0         2.0
#> 115  virginica          5.8         2.8          5.1         2.4
#> 116  virginica          6.4         3.2          5.3         2.3
#> 117  virginica          6.5         3.0          5.5         1.8
#> 118  virginica          7.7         3.8          6.7         2.2
#> 119  virginica          7.7         2.6          6.9         2.3
#> 120  virginica          6.0         2.2          5.0         1.5
#> 121  virginica          6.9         3.2          5.7         2.3
#> 122  virginica          5.6         2.8          4.9         2.0
#> 123  virginica          7.7         2.8          6.7         2.0
#> 124  virginica          6.3         2.7          4.9         1.8
#> 125  virginica          6.7         3.3          5.7         2.1
#> 126  virginica          7.2         3.2          6.0         1.8
#> 127  virginica          6.2         2.8          4.8         1.8
#> 128  virginica          6.1         3.0          4.9         1.8
#> 129  virginica          6.4         2.8          5.6         2.1
#> 130  virginica          7.2         3.0          5.8         1.6
#> 131  virginica          7.4         2.8          6.1         1.9
#> 132  virginica          7.9         3.8          6.4         2.0
#> 133  virginica          6.4         2.8          5.6         2.2
#> 134  virginica          6.3         2.8          5.1         1.5
#> 135  virginica          6.1         2.6          5.6         1.4
#> 136  virginica          7.7         3.0          6.1         2.3
#> 137  virginica          6.3         3.4          5.6         2.4
#> 138  virginica          6.4         3.1          5.5         1.8
#> 139  virginica          6.0         3.0          4.8         1.8
#> 140  virginica          6.9         3.1          5.4         2.1
#> 141  virginica          6.7         3.1          5.6         2.4
#> 142  virginica          6.9         3.1          5.1         2.3
#> 143  virginica          5.8         2.7          5.1         1.9
#> 144  virginica          6.8         3.2          5.9         2.3
#> 145  virginica          6.7         3.3          5.7         2.5
#> 146  virginica          6.7         3.0          5.2         2.3
#> 147  virginica          6.3         2.5          5.0         1.9
#> 148  virginica          6.5         3.0          5.2         2.0
#> 149  virginica          6.2         3.4          5.4         2.3
#> 150  virginica          5.9         3.0          5.1         1.8