Apply a data dictionary to a dataset

Applies a data dictionary to a data structure, creating a labelled dataset. All previous attributes will be preserved. For factors, the attribute 'levels' will be transformed into attribute 'labels' and values will be recast into appropriate datatypes.

data_dict_apply(dataset, data_dict = NULL)

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

Value

A tibble identifying the dataset with the data dictionary applied to each variable as attributes.

Details

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

See also

Examples

{

# use DEMO_files provided by the package

dataset <- DEMO_files$dataset_MELBOURNE_1
data_dict <- as_data_dict_mlstr(DEMO_files$dd_MELBOURNE_1_format_maelstrom)
data_dict_apply(dataset, data_dict)

}
#> # A tibble: 19 × 4
#>        id Gender       BMI age                        
#>     <dbl> <dbl+lbl>  <dbl> <dbl+lbl>                  
#>  1 377943 2 [Female]  22.1   52                       
#>  2 497013 1 [Male]    18.5   49                       
#>  3 927676 1 [Male]    24.6   43                       
#>  4 995667 2 [Female]  15.7   59                       
#>  5  21829 2 [Female]  18.6   40                       
#>  6 209432 1 [Male]    15.9   47                       
#>  7 272983 2 [Female]  28.7 -888 [don't want to answer]
#>  8 580632 2 [Female]  19.7   53                       
#>  9 304624 2 [Female]  NA     35                       
#> 10 637551 1 [Male]    21.3   40                       
#> 11 279817 1 [Male]    17.2   41                       
#> 12 235415 2 [Female]  21.7   34                       
#> 13 373673 2 [Female]  14.7   48                       
#> 14 485098 2 [Female]  24.4   43                       
#> 15 299427 1 [Male]    NA   -888 [don't want to answer]
#> 16 854073 1 [Male]    24.2   41                       
#> 17 197666 2 [Female]  17.7   33                       
#> 18 130327 2 [Female]  19.2   57                       
#> 19 220050 2 [Female]  18.8   50