Generate a quality assessment report of a data dictionary

Assesses the content and structure of a data dictionary and reports potential issues to facilitate the assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. This report is compatible with Excel and can be exported as an Excel spredsheet.

data_dict_evaluate(data_dict, taxonomy = NULL, as_data_dict_mlstr = TRUE)

Arguments

data_dict

A list of tibble(s) representing meta data to be evaluated.

taxonomy

A tibble identifying the scheme used for variables classification as a tibble.

as_data_dict_mlstr

Whether the output data dictionary has a simple data dictionary structure or not (meaning has a Maelstrom data dictionary structure, compatible with Maelstrom Research ecosystem, including Opal). TRUE by default.

Value

A list of tibbles of report for one data dictionary.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

Examples

{

# use DEMO_files provided by the package

data_dict <- DEMO_files$`dd_TOKYO_format_maelstrom_tagged - ERROR`
data_dict_evaluate(data_dict)

}
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> $`Data dictionary summary`
#> # A tibble: 14 × 20
#>    index name    `label:en` valueType `Categories::table` `Categories::label:en`
#>    <int> <chr>   <chr>      <chr>     <chr>               <chr>                 
#>  1     1 part_id id of the… NA         NA                  NA                   
#>  2     2 gndr    gndr       boolean   "Male = DEMO ; \nF… "Male = Male ; \nFema…
#>  3     3 height  height     decimal   "SKIP PATTERN = DE… "SKIP PATTERN = SKIP …
#>  4     4 weight… weight_ms  integer   "-88 = DEMO ; \n-9… "-88 = Don’t want to …
#>  5     5 weight… weight_dc  integer    NA                  NA                   
#>  6     6 weight… weight_dc  integr     NA                  NA                   
#>  7     7 dob     NA         date       NA                  NA                   
#>  8     8 prg_evr prg_ever   integer    NA                  NA                   
#>  9     9 empty   empty      integer    NA                  NA                   
#> 10    10 opente… opentext   character  NA                  NA                   
#> 11    11 opente… NA         NA         NA                  NA                   
#> 12    12 NA      no name    integer   "6 = DEMO"          "6 = Inaccurate"      
#> 13    NA prg_ev… NA         NA        "0 = DEMO ; \n1 = … "0 = never pregnant ;…
#> 14    NA weight… NA         NA        "-99 = DEMO"        "-99 = Don’t want to …
#> # ℹ 14 more variables: `Categories::missing` <chr>, table <chr>,
#> #   `description:en` <chr>, `copy units` <chr>, unit <chr>,
#> #   `Mlstr_additional::Source` <chr>, `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::1` <chr>, `Mlstr_area::1.term` <chr>,
#> #   `Mlstr_area::1.scale` <chr>, `Mlstr_area::2` <chr>,
#> #   `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <chr>,
#> #   `Mlstr_area::3.term` <chr>
#> 
#> $`Data dictionary assessment`
#> # A tibble: 19 × 4
#>    sheet      name_col           name_var       `Quality assessment comment`    
#>    <chr>      <chr>              <chr>          <chr>                           
#>  1 Variables  Mlstr_area::3      NA             [INFO] - Empty column           
#>  2 Variables  Mlstr_area::3.term NA             [INFO] - Empty column           
#>  3 Variables  copy units         NA             [INFO] - Possible duplicated co…
#>  4 Variables  label:en           dob            [INFO] - The column `label:en` …
#>  5 Variables  label:en           opentext       [INFO] - The column `label:en` …
#>  6 Variables  name               opentext       [ERR] - duplicated variable name
#>  7 Variables  name               row number: 12 [ERR] - missing variable name   
#>  8 Variables  name               weight dc      [INFO] - Incompatible variable …
#>  9 Variables  unit               NA             [INFO] - Possible duplicated co…
#> 10 Variables  valueType          gndr           [ERR] - valueType conflict in '…
#> 11 Variables  valueType          height         [ERR] - valueType conflict in '…
#> 12 Variables  valueType          opentext       [ERR] - Incompatible valueType …
#> 13 Variables  valueType          weight_dc      [ERR] - Incompatible valueType …
#> 14 Categories label:en           1              [INFO] - The column `label:en` …
#> 15 Categories variable           prg_ever       [ERR] - Categories not present …
#> 16 Categories variable           prg_ever       [ERR] - In 'name', the value ca…
#> 17 Categories variable           prg_ever       [ERR] - Category names not uniq…
#> 18 Categories variable           weight_sm      [ERR] - Categories not present …
#> 19 Categories variable           NA             [ERR] - In 'variable', the valu…
#>