Transform single-row category information to multiple rows and add as element

Expands data dictionary column(s) in a element (the parameter 'from'), into another element (the parameter 'to'). If the element 'from' contains any column starting with 'prefix', (xx,yy), these columns will be added as 'xx' and 'yy' in the element identified by 'to'. This tibble will be created if necessary, and columns will be added, from left to right. (unique names will be generated if necessary). Separator of each element is the following structure : 'name = xx1 ; name = xx2'. This function is mainly used to expand the column(s) 'Categories::xx' in "Variables" to "Categories" element with column(s) xx. This function is the reversed operation of data_dict_collapse()

data_dict_expand(
  data_dict,
  from = "Variables",
  name_prefix = "Categories::",
  to = "Categories"
)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed. Automatically generated if not provided.

from

Symbol identifying the name of the element (tibble) to take column(s) from. Default is 'Variables'.

name_prefix

Character string of the prefix of columns of interest. This prefix will be used to select columns, and to rename them in the 'to' element. Default is 'Categories::'.

to

Symbol identifying the name of the element (tibble) to create column(s) to. Default is 'Categories'.

Value

A list of tibble(s) identifying a data dictionary.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

See also

Examples

{

# use DEMO_files provided by the package

data_dict <- DEMO_files$`dd_PARIS_format_flatten`
data_dict_expand(data_dict)

}
#> $Variables
#> # A tibble: 7 × 18
#>   name     `label:fr`    `description:fr` valueType unit  Mlstr_additional::So…¹
#>   <chr>    <chr>         <chr>            <chr>     <chr> <chr>                 
#> 1 ID       id            id du participa… text      NA    Questionnaire         
#> 2 SEX      Sexe          Sexe du partici… integer   NA    Questionnaire         
#> 3 BMI      IMC           Indice de Masse… decimal   kg/m… Physical_measure      
#> 4 AGE      Age           Age du Particip… integer   years Physical_measure      
#> 5 SMO      Fumeur        Fumeur regulier  integer   NA    Physical_measure      
#> 6 SMO_QTY  fum_qte_sema… nombre de cigar… integer   ciga… Questionnaire         
#> 7 PRG_EVER enceinte_occ… Si la personne … integer   NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 12 more variables: `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::Administrative_information` <chr>,
#> #   `Mlstr_area::Sociodemographic_economic_characteristics` <chr>,
#> #   `Mlstr_area::Physical_measures` <chr>, `Mlstr_area::Reproduction` <chr>,
#> #   `Mlstr_area::Lifestyle_behaviours` <chr>,
#> #   `Mlstr_cogscale::Cog_scale` <chr>, `Mlstr_habits::Tobacco` <chr>, …
#> 
#> $Categories
#> # A tibble: 9 × 3
#>   name  variable label                                        
#>   <chr> <chr>    <chr>                                        
#> 1 0     SEX      Homme                                        
#> 2 1     SEX      Femme                                        
#> 3 0     SMO      Non-fumeur                                   
#> 4 1     SMO      Fumeur (cigarette ; cigare)                  
#> 5 -8    SMO_QTY  SKIP PATTERN                                 
#> 6 0     PRG_EVER Jamais enceinte                              
#> 7 1     PRG_EVER Enceinte au moins une fois (nb grossesse = 0)
#> 8 9     PRG_EVER NSP                                          
#> 9 -8    PRG_EVER SKIP PATTERN                                 
#> 
#> attr(,"madshapR::class")
#> [1] "data_dict_structure"