Transform column(s) of a data dictionary from wide format to long format

Transforms column(s) of a data dictionary from wide format to long format. If a taxonomy is provided, the corresponding columns in the data dictionary will be converted to a standardized format with fewer columns. This operation is equivalent to performing a tidyr::pivot_longer() on these columns following the taxonomy structure provided. Variable names in the data dictionary must be unique.

data_dict_pivot_longer(data_dict, taxonomy = NULL)

Arguments

data_dict

A list of tibble(s) representing meta data to be transformed.

taxonomy

A tibble identifying the scheme used for variables classification.

Value

A list of tibble(s) identifying a data dictionary.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# use DEMO_files provided by the package

taxonomy <- DEMO_files$taxonomy_opal_mlstr
data_dict <- DEMO_files$dd_TOKYO_format_opal_tagged
data_dict_pivot_longer(data_dict,taxonomy)

}
#> $Variables
#> # A tibble: 9 × 14
#>   name      `label:en`            `description:en`  valueType unit  other_format
#>   <chr>     <chr>                 <chr>             <chr>     <chr> <lgl>       
#> 1 part_id   id of the participant id of the partic… text      NA    NA          
#> 2 gndr      gndr                  gender of the pa… text      NA    NA          
#> 3 height    height                height of the pa… integer   cm    NA          
#> 4 weight_ms weight_ms             weight of the pa… integer   kg    NA          
#> 5 weight_dc weight_dc             weight of the pa… decimal   kg    NA          
#> 6 dob       dob                   date of birth of… date      years NA          
#> 7 prg_ever  prg_ever              whether the part… integer   NA    NA          
#> 8 empty     empty                 empty column      integer   NA    NA          
#> 9 opentext  opentext              open text         text      NA    NA          
#> # ℹ 8 more variables: `Mlstr_additional::1` <chr>,
#> #   `Mlstr_additional::1.term` <chr>, `Mlstr_additional::2` <chr>,
#> #   `Mlstr_additional::2.term` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::2` <chr>,
#> #   `Mlstr_area::2.term` <chr>
#> 
#> $Categories
#> # A tibble: 9 × 4
#>   variable name   `label:en`            missing
#>   <chr>    <chr>  <chr>                   <dbl>
#> 1 gndr     Male   Male                        0
#> 2 gndr     Female Female                      0
#> 3 gndr     -77    Don’t want to answer        1
#> 4 prg_ever 0      never pregnant              0
#> 5 prg_ever 1      pregnant once or more       0
#> 6 prg_ever 2      currently pregnant          0
#> 7 prg_ever 8      Don’t want to answer        1
#> 8 prg_ever 9      Don’t know                  1
#> 9 prg_ever -7     not applicable              1
#>