Apply data dictionary category labels to the associated dataset variables

Applies category labels declared in a data dictionary to the associated columns (variables) in the dataset.

dataset_cat_as_labels(dataset, data_dict = NULL, col_names = names(dataset))

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset (to be generated).

col_names

A character string specifying the name(s) of the column(s) which refer to existing column(s) in the dataset. The column(s) can be named or indicated by position.

Value

A tibble identifying a dataset.

Details

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

Examples

{

# use DEMO_files provided by the package

dataset <- DEMO_files$dataset_TOKYO
data_dict <- as_data_dict_mlstr(DEMO_files$dd_TOKYO_format_maelstrom_tagged)
dataset <- data_dict_apply(dataset,data_dict)
dataset_cat_as_labels(dataset)

}
#> # A tibble: 50 × 9
#>    part_id gndr       height weight_ms  weight_dc dob   prg_ever  empty opentext
#>    <chr>   <chr+lbl>   <dbl> <chr+lbl>      <dbl> <chr> <chr+lbl> <lgl> <chr>   
#>  1 ID001   Male [Mal…    191 63                NA 3/22… not… [-7] NA    All chi…
#>  2 ID002   Fem… [Fem…    176 NA                65 8/15… neve… [0] NA    grow up…
#>  3 ID003   Fem… [Fem…    154 NA               121 12/1… curr… [2] NA    years o…
#>  4 ID004   Fem… [Fem…    167 Don… [-88]        NA 6/13… preg… [1] NA    flower …
#>  5 ID005   Fem… [Fem…    185 NA                45 12/1… Don’… [8] NA    rather …
#>  6 ID006   Don… [-77]    171 57                NA 3/31… not… [-7] NA    cried, …
#>  7 ID007   Fem… [Fem…    185 NA                58 4/19… Don’… [9] NA    that pa…
#>  8 ID008   Fem… [Fem…    171 NA                59 NA    curr… [2] NA    that sh…
#>  9 ID009   Don… [-77]    169 52                NA 3/14… not… [-7] NA    beginni…
#> 10 ID010   Male [Mal…    179 NA                62 10/1… not… [-7] NA    All chi…
#> # ℹ 40 more rows