Generate a report and summary of a dataset

Assesses and summarizes the content and structure of a dataset and data dictionary and reports potential issues to facilitate the assessment of input. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. The summary provides additional information about variable distributions and descriptive statistics. This report is compatible with Excel and can be exported as an Excel spreadsheet.

dataset_summarize(
  dataset,
  data_dict = data_dict_extract(dataset),
  group_by = NULL,
  taxonomy = NULL,
  dataset_name = .dataset_name,
  valueType_guess = FALSE,
  .dataset_name = NULL
)

Arguments

dataset

A data frame identifying the input dataset observations associated to its data dictionary.

data_dict

A list of data frame(s) representing meta data of an associated dataset. Automatically generated if not provided.

group_by

A character string identifying the column in the dataset to use as a grouping variable. Visual elements will be grouped by this column.

taxonomy

An optional data frame identifying a variable classification schema.

dataset_name

A character string specifying the name of the dataset (internally used in the function dossier_evaluate()).

valueType_guess

Whether the output should be generated based on more precise valueType inferred from the data. FALSE by default (will use the valueType declared).

.dataset_name

Value

A list of data frames of report for one data dictionary.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. An identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

A taxonomy is a classification schema that can be defined for variable attributes. A taxonomy is usually extracted from an Opal environment, and a taxonomy object is a data frame that must contain at least the columns taxonomy, vocabulary, and terms. Additional details about Opal taxonomies are available online.

The valueType is a declared property of a variable that is required in certain functions to determine handling of the variables. Specifically, valueType refers to the OBiBa data type of a variable. The valueType is specified in a data dictionary in a column 'valueType' and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'. The full list of OBiBa valueType possibilities and their correspondence with R data types are available using valueType_list.

See also

Examples

{

# use madshapR_DEMO provided by the package
library(dplyr)

#' ###### Example : any data frame can be summarized
dataset <- iris['Sepal.Width']
dataset_summarize(dataset)
 
}
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>     The dataset contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarize information for all variables
#>     Summarize information for numerical variables
#>     Summarize information for text variables
#>     Summarize information for date variables
#>     Summarize information for datetime variables
#>     Summarize information for categorical variables
#>     Summarize global information (Overview)
#>     Generate report
#> $Overview
#> # A tibble: 15 × 2
#>    `Quality control of dataset`                           `(all)`     
#>    <chr>                                                  <chr>       
#>  1 "Date"                                                 "2023-12-05"
#>  2 "Name of the dataset"                                  "dataset"   
#>  3 "    Identifier Variable"                              ""          
#>  4 "    Variables"                                        " "         
#>  5 "        Total number of variables (incl. identifier)" "1"         
#>  6 "        Total number of empty columns"                "0"         
#>  7 "    Data type in dictionary (valueType)"              " "         
#>  8 "        Nb. text variables"                           "0"         
#>  9 "        Nb. date variables"                           "0"         
#> 10 "        Nb. datetime variables"                       "0"         
#> 11 "        Nb. numerical variables"                      "1"         
#> 12 "        Nb. categorical variables"                    "0"         
#> 13 "    Rows"                                             "(all)"     
#> 14 "        Total number of observations"                 "150"       
#> 15 "        Nb. distinct observations"                    "150"       
#> 
#> $`Data dictionary assessment`
#> # A tibble: 2 × 5
#>   sheet     col_name name_var `Quality assessment comment`         value       
#>   <chr>     <chr>    <chr>    <chr>                                <chr>       
#> 1 Variables label    NA       [INFO] - Possible duplicated columns name ; label
#> 2 Variables name     NA       [INFO] - Possible duplicated columns name ; label
#> 
#> $`Variables summary (all)`
#> # A tibble: 1 × 13
#>   `index in data dict.` name        `Quality assessment comment` label      
#>                   <int> <chr>       <chr>                        <chr>      
#> 1                     1 Sepal.Width NA                           Sepal.Width
#> # ℹ 9 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>
#> 
#> $`Numerical variable summary`
#> # A tibble: 1 × 22
#>   `index in data dict.` name        `Quality assessment comment` label      
#>                   <int> <chr>       <chr>                        <chr>      
#> 1                     1 Sepal.Width NA                           Sepal.Width
#> # ℹ 18 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, …
#>