Generate a web-based bookdown visual report of a dataset

Generates a visual report for a dataset in an HTML bookdown document. The report provides figures and descriptive statistics for each variable to facilitate the assessment of input data. Statistics and figures are generated according to variable data type. The report can be used to help assess data structure, coherence across elements, and taxonomy or data dictionary formats. The summaries and figures provide additional information about variable distributions and descriptive statistics. The charts and tables are produced based on their data type. The variable can be grouped using group_by parameter, which is a (categorical) column in the dataset. The user may need to use as.factor() in this context. To fasten the process (and allow recycling object in a workflow) the user can feed the function with a .summary_var, which is the output of the function dataset_summarize() of the column(s) col and group_by. The summary must have the same parameters to operate.

dataset_visualize(
  dataset = tibble(id = as.character()),
  data_dict = data_dict_extract(dataset),
  group_by = NULL,
  to,
  taxonomy = NULL,
  .summary_var = NULL,
  .dataset_name = NULL,
  .keep_files = TRUE
)

Arguments

dataset

A tibble identifying the input dataset observations associated to its data dictionary.

data_dict

A list of tibble(s) representing meta data of an associated dataset. Automatically generated if not provided.

group_by

A character string of one column in the dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

to

A character string identifying the folder path where the bookdown report will be saved.

taxonomy

A tibble identifying the scheme used for variables classification.

.summary_var

A list which is the summary of the variables.

.dataset_name

A character string specifying the name of the dataset (used for internal processes and programming).

.keep_files

whether to keep the R-markdown files. TRUE by default. (used for internal processes and programming)

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'docs/index.html'. Or use open_visual_report()

Details

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame-like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# use DEMO_files provided by the package
library(dplyr)
library(fabR)  # silently_run

###### Example : Combine functions and summarise datasets.
data_dict <- as_data_dict_mlstr(DEMO_files$dd_TOKYO_format_maelstrom_tagged)
dataset <-
  DEMO_files$dataset_TOKYO %>%
  valueType_adjust(from = data_dict) %>%
  data_dict_apply(data_dict)

# Silencing the process, remove silently_run() to see the full process in 
# the console. 
to <- tempdir()
silently_run(dataset_visualize(dataset,group_by = gndr,to = to))
  
# To open the file in browser, open 'to/docs/index.html'. 
# Or use [open_visual_report(to)]

}
#>   |                                                           |                                                   |   0%  |                                                           |.                                                  |   1%                     |                                                           |.                                                  |   2% (unnamed-chunk-1)   |                                                           |..                                                 |   4%                     |                                                           |..                                                 |   5% (unnamed-chunk-2)   |                                                           |...                                                |   6%                     |                                                           |....                                               |   7% (unnamed-chunk-3)   |                                                           |....                                               |   8%                     |                                                           |.....                                              |   9% (unnamed-chunk-4)   |                                                           |.....                                              |  11%                     |                                                           |......                                             |  12% (unnamed-chunk-5)   |                                                           |.......                                            |  13%                     |                                                           |.......                                            |  14% (unnamed-chunk-6)   |                                                           |........                                           |  15%                     |                                                           |........                                           |  16% (figures-plot12-1)  |                                                           |.........                                          |  18%                     |                                                           |..........                                         |  19% (unnamed-chunk-7)   |                                                           |..........                                         |  20%                     |                                                           |...........                                        |  21% (unnamed-chunk-8)   |                                                           |...........                                        |  22%                     |                                                           |............                                       |  24% (unnamed-chunk-9)   |                                                           |.............                                      |  25%                     |                                                           |.............                                      |  26% (unnamed-chunk-10)  |                                                           |..............                                     |  27%                     |                                                           |..............                                     |  28% (figures-plot12-2)  |                                                           |...............                                    |  29%                     |                                                           |................                                   |  31% (unnamed-chunk-11)  |                                                           |................                                   |  32%                     |                                                           |.................                                  |  33% (unnamed-chunk-12)  |                                                           |.................                                  |  34%                     |                                                           |..................                                 |  35% (unnamed-chunk-13)  |                                                           |...................                                |  36%                     |                                                           |...................                                |  38% (figures-plot12-3)  |                                                           |....................                               |  39%                     |                                                           |....................                               |  40% (unnamed-chunk-14)  |                                                           |.....................                              |  41%                     |                                                           |......................                             |  42% (unnamed-chunk-15)  |                                                           |......................                             |  44%                     |                                                           |.......................                            |  45% (unnamed-chunk-16)  |                                                           |.......................                            |  46%                     |                                                           |........................                           |  47% (unnamed-chunk-17)  |                                                           |.........................                          |  48%                     |                                                           |.........................                          |  49% (figures-plot12-4)  |                                                           |..........................                         |  51%                     |                                                           |..........................                         |  52% (unnamed-chunk-18)  |                                                           |...........................                        |  53%                     |                                                           |............................                       |  54% (unnamed-chunk-19)  |                                                           |............................                       |  55%                     |                                                           |.............................                      |  56% (unnamed-chunk-20)  |                                                           |.............................                      |  58%                     |                                                           |..............................                     |  59% (figures-plot12-5)  |                                                           |...............................                    |  60%                     |                                                           |...............................                    |  61% (unnamed-chunk-21)  |                                                           |................................                   |  62%                     |                                                           |................................                   |  64% (unnamed-chunk-22)  |                                                           |.................................                  |  65%                     |                                                           |..................................                 |  66% (unnamed-chunk-23)  |                                                           |..................................                 |  67%                     |                                                           |...................................                |  68% (figures-plot12-6)  |                                                           |...................................                |  69%                     |                                                           |....................................               |  71% (unnamed-chunk-24)  |                                                           |.....................................              |  72%                     |                                                           |.....................................              |  73% (unnamed-chunk-25)  |                                                           |......................................             |  74%                     |                                                           |......................................             |  75% (unnamed-chunk-26)  |                                                           |.......................................            |  76%                     |                                                           |........................................           |  78% (unnamed-chunk-27)  |                                                           |........................................           |  79%                     |                                                           |.........................................          |  80% (figures-plot12-7)  |                                                           |.........................................          |  81%                     |                                                           |..........................................         |  82% (unnamed-chunk-28)  |                                                           |...........................................        |  84%                     |                                                           |...........................................        |  85% (unnamed-chunk-29)  |                                                           |............................................       |  86%                     |                                                           |............................................       |  87% (unnamed-chunk-30)  |                                                           |.............................................      |  88%                     |                                                           |..............................................     |  89% (figures-plot12-8)  |                                                           |..............................................     |  91%                     |                                                           |...............................................    |  92% (unnamed-chunk-31)  |                                                           |...............................................    |  93%                     |                                                           |................................................   |  94% (unnamed-chunk-32)  |                                                           |.................................................  |  95%                     |                                                           |.................................................  |  96% (unnamed-chunk-33)  |                                                           |.................................................. |  98%                     |                                                           |.................................................. |  99% (figures-plot12-9)  |                                                           |...................................................| 100%                                                                                                                                
#> "C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/pandoc" +RTS -K512m -RTS bookdownproj.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output bookdownproj.html --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\rmarkdown\lua\custom-environment.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\latex-div.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\anchor-sections.lua" --metadata-file "C:\Users\guill\AppData\Local\Temp\RtmpiwmlpV\file4744629e579f" --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\templates\gitbook.html" --highlight-style pygments --number-sections --css style.css --mathjax --include-in-header "C:\Users\guill\AppData\Local\Temp\RtmpiwmlpV\rmarkdown-str47444e593bf6.html"