Generate an Excel spreadsheet report of a dossier (dataset list)

Generates an Excel spreadsheet report for a dataset list (or dossier) showing descriptive statistics for each variable to facilitate the assessment of input dataset. Statistics are generated according to their valueType. This report can be used to assist the user in the assessment of the dataset structure, fields investigation (mandatory or not), coherence across elements and taxonomy, or standard evaluation. The summary associated provides dataset composition, with observation distribution and descriptive statistics.

dossier_summarize(dossier, taxonomy = NULL, valueType_guess = TRUE)

Arguments

dossier

List of tibble, each of them being datasets.

taxonomy

A data-frame or data-frame extension (e.g. a tibble), identifying the scheme used for variables classification as a tibble.

valueType_guess

Whether the output should include a more accurate valueType that could be applied to the dataset. TRUE by default.

Value

A list of tibbles of report for each dataset.

Details

A dossier must be a named list containing at least one data-frame or data-frame extension (e.g. a tibble), each of them being datasets. The name of each tibble will be use as the reference name of the dataset.

A data dictionary-like structure must be a list of at least one or two data-frame or data-frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

Examples

{

# use DEMO_files provided by the package
library(dplyr)

###### Example 1: Combine functions and summarise datasets.
dossier <- 
  DEMO_files[stringr::str_detect(names(DEMO_files),"dataset_MELBOURNE")] %>%
  lapply(valueType_self_adjust)

dossier_summarize(dossier)

}
#> - DOSSIER SUMMARY: -----------------------------------------------------
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset_MELBOURNE_1 --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>     The dataset contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset_MELBOURNE_1 --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for numerical variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset_MELBOURNE_2 --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>     The dataset contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset_MELBOURNE_2 --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for numerical variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#> $dataset_MELBOURNE_1
#> $dataset_MELBOURNE_1$Overview
#> # A tibble: 13 × 2
#>    `Quality control of dataset`              `(all)`              
#>    <chr>                                     <chr>                
#>  1 "Date"                                    "2023-05-08"         
#>  2 "Name of the dataset"                     "dataset_MELBOURNE_1"
#>  3 "    Identifier Variable"                 ""                   
#>  4 "    Variables"                           " "                  
#>  5 "        Total number of variables"       "4"                  
#>  6 "    Data type in dictionary (valueType)" " "                  
#>  7 "        Nb. text variables"              "0"                  
#>  8 "        Nb. date variables"              "0"                  
#>  9 "        Nb. numerical variables"         "4"                  
#> 10 "        Nb. categorical variables"       "0"                  
#> 11 "    Rows"                                "(all)"              
#> 12 "        Total number of observations"    "19"                 
#> 13 "        Nb. distinct observations"       "19"                 
#> 
#> $dataset_MELBOURNE_1$`Data dictionary assessment`
#> # A tibble: 2 × 4
#>   sheet     name_col name_var `Quality assessment comment`                      
#>   <chr>     <chr>    <chr>    <chr>                                             
#> 1 Variables label    NA       [INFO] - Possible duplicated columns: name ; label
#> 2 Variables name     NA       [INFO] - Possible duplicated columns: name ; label
#> 
#> $dataset_MELBOURNE_1$`Variables summary (all)`
#> # A tibble: 4 × 15
#>   `index in data dict.` name   `Quality assessment comment`         label 
#>                   <int> <chr>  <chr>                                <chr> 
#> 1                     1 id     [INFO] - All observations are unique id    
#> 2                     2 Gender NA                                   Gender
#> 3                     3 BMI    NA                                   BMI   
#> 4                     4 age    NA                                   age   
#> # ℹ 11 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>
#> 
#> $dataset_MELBOURNE_1$`Numerical variable summary`
#> # A tibble: 4 × 22
#>   `index in data dict.` name   `Quality assessment comment`         label 
#>                   <int> <chr>  <chr>                                <chr> 
#> 1                     1 id     [INFO] - All observations are unique id    
#> 2                     2 Gender NA                                   Gender
#> 3                     3 BMI    NA                                   BMI   
#> 4                     4 age    NA                                   age   
#> # ℹ 18 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, …
#> 
#> 
#> $dataset_MELBOURNE_2
#> $dataset_MELBOURNE_2$Overview
#> # A tibble: 13 × 2
#>    `Quality control of dataset`              `(all)`              
#>    <chr>                                     <chr>                
#>  1 "Date"                                    "2023-05-08"         
#>  2 "Name of the dataset"                     "dataset_MELBOURNE_2"
#>  3 "    Identifier Variable"                 ""                   
#>  4 "    Variables"                           " "                  
#>  5 "        Total number of variables"       "3"                  
#>  6 "    Data type in dictionary (valueType)" " "                  
#>  7 "        Nb. text variables"              "0"                  
#>  8 "        Nb. date variables"              "0"                  
#>  9 "        Nb. numerical variables"         "3"                  
#> 10 "        Nb. categorical variables"       "0"                  
#> 11 "    Rows"                                "(all)"              
#> 12 "        Total number of observations"    "19"                 
#> 13 "        Nb. distinct observations"       "19"                 
#> 
#> $dataset_MELBOURNE_2$`Data dictionary assessment`
#> # A tibble: 2 × 4
#>   sheet     name_col name_var `Quality assessment comment`                      
#>   <chr>     <chr>    <chr>    <chr>                                             
#> 1 Variables label    NA       [INFO] - Possible duplicated columns: name ; label
#> 2 Variables name     NA       [INFO] - Possible duplicated columns: name ; label
#> 
#> $dataset_MELBOURNE_2$`Variables summary (all)`
#> # A tibble: 3 × 15
#>   `index in data dict.` name       `Quality assessment comment`         label   
#>                   <int> <chr>      <chr>                                <chr>   
#> 1                     1 id         [INFO] - All observations are unique id      
#> 2                     2 smo_status NA                                   smo_sta…
#> 3                     3 prg_curr   NA                                   prg_curr
#> # ℹ 11 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>
#> 
#> $dataset_MELBOURNE_2$`Numerical variable summary`
#> # A tibble: 3 × 22
#>   `index in data dict.` name       `Quality assessment comment`         label   
#>                   <int> <chr>      <chr>                                <chr>   
#> 1                     1 id         [INFO] - All observations are unique id      
#> 2                     2 smo_status NA                                   smo_sta…
#> 3                     3 prg_curr   NA                                   prg_curr
#> # ℹ 18 more variables: `Data Dictionary valueType` <chr>,
#> #   `Estimated dataset valueType` <chr>, `Actual dataset valueType` <chr>,
#> #   `Categorical variable` <chr>, `Categories in data dictionary` <chr>,
#> #   `Total number of observations` <dbl>, `Nb. distinct values` <int>,
#> #   `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, …
#> 
#>