Provide descriptive statistics for variables of type 'datetime' in a dataset

Summarizes (in a data frame) the columns of type 'datetime' in a dataset and its data dictionary (if any). The summary provides information about quality, type, composition, and descriptive statistics of variables. Statistics are generated by valueType.

summary_variables_datetime(
  dataset_preprocess,
  dataset = NULL,
  data_dict = NULL
)

Arguments

dataset_preprocess

A data frame which provides summary of the variables (for internal processes and programming).

dataset

A dataset object.

data_dict

A list of data frame(s) representing metadata of the input dataset. Automatically generated if not provided.

Value

A data frame providing statistical description of 'datetime' variables present in a dataset.

Details

A data dictionary contains the list of variables in a dataset and metadata about the variables and can be associated with a dataset. A data dictionary object is a list of data frame(s) named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the data frame 'Variables' must contain at least the name column, with all unique and non-missing entries, and the data frame 'Categories' must contain at least the variable and name columns, with unique combination of variable and name.

A dataset is a data table containing variables. A dataset object is a data frame and can be associated with a data dictionary. If no data dictionary is provided with a dataset, a minimum workable data dictionary will be generated as needed within relevant functions. Identifier variable(s) for indexing can be specified by the user. The id values must be non-missing and will be used in functions that require it. If no identifier variable is specified, indexing is handled automatically by the function.

Examples

{
   
library(dplyr)
library(lubridate)
library(fabR)

# use madshapR_examples provided by the package
dataset <- madshapR_examples$`dataset_example` %>%
  mutate(dob = as_datetime(as_any_date(dob))) %>%
  select(dob)

dataset_preprocess <- dataset_preprocess(dataset)
summary_prep <- summary_variables_datetime(
  dataset_preprocess = dataset_preprocess[[1]])
head(summary_prep)

}
#> 
#> Attaching package: 'lubridate'
#> The following objects are masked from 'package:base':
#> 
#>     date, intersect, setdiff, union
#> # A tibble: 1 × 3
#>   `Variable name` `Most common values` `Least common values`                    
#>   <chr>           <chr>                <chr>                                    
#> 1 dob             1962-06-14           1960-09-22 ; 1961-02-24 ; 1961-04-20 ; 1…