Attribute the valueType from a data dictionary to a dataset, or vice versa

Takes the valueType of the input (from) and attributes it to the output (to). The parameters 'from' and 'to' can be either a dataset or a data dictionary. Depending on the input provided, the valueType replaced is either in the 'valueType' column of a data dictionary or cast to a column in a dataset. If 'to' is not provided, the function calls valueType_self_adjust() instead. The possible values returned are 'date', 'boolean', 'integer', 'decimal', and text'.

valueType_adjust(from, to = NULL)

Arguments

from

Object to be adjusted. Can be either a dataset or a data dictionary.

to

Object to be adjusted. Can be either a dataset or a data dictionary. NULL by default.

Value

Either a tibble, identifying the dataset, or a list of tibble(s) identifying a data dictionary, depending which is 'to'.

Details

A data dictionary contains metadata about variables and can be associated with a dataset. It must be a list of data frame-like objects with elements named 'Variables' (required) and 'Categories' (if any). To be usable in any function, the 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns. To be considered as a minimum workable data dictionary, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A dataset must be a data frame-like object and can be associated with a data dictionary. If no data dictionary is provided, a minimum workable data dictionary will be generated as needed by relevant functions. An identifier id column for sorting can be specified by the user. If specified, the id values must be non-missing and will be used in functions that require it. If no identifier column is specified, indexing is handled automatically by the function.

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondence with R data types are available using valueType_list.

See also

Examples

{

# use DEMO_files provided by the package
library(dplyr)

dataset <- DEMO_files$dataset_TOKYO[c(1:4),'prg_ever']
data_dict <-
  DEMO_files$dd_TOKYO_format_maelstrom %>%
  data_dict_filter(filter_var = 'name == "prg_ever"') %>%
  as_data_dict_mlstr()
valueType_adjust(from = dataset,to = data_dict)

}
#> $Variables
#> # A tibble: 1 × 6
#>   name     `label:en` valueType index `description:en`                    typeof
#>   <chr>    <chr>      <chr>     <chr> <chr>                               <chr> 
#> 1 prg_ever prg_ever   decimal   8     whether the participant has ever b… double
#> 
#> $Categories
#> # A tibble: 6 × 4
#>   variable name  `label:en`            missing
#>   <chr>    <chr> <chr>                 <lgl>  
#> 1 prg_ever 0     never pregnant        FALSE  
#> 2 prg_ever 1     pregnant once or more FALSE  
#> 3 prg_ever 2     currently pregnant    FALSE  
#> 4 prg_ever 8     Don’t want to answer  TRUE   
#> 5 prg_ever 9     Don’t know            TRUE   
#> 6 prg_ever -7    not applicable        TRUE   
#> 
#> attr(,"madshapR::class")
#> [1] "data_dict_mlstr"