Search and replace the valueType of any data dictionary or dataset

Using the function valueType_guess(), this function provides the first possible valueType, by trying to assign the object to 'boolean', then 'integer', then 'decimal', then 'date'. (If all fails, 'text', by default), then either replace the valueType column of a data dictionary of casts a column in the dataset. This depending on the input provided.

valueType_self_adjust(...)

Arguments

...

R object that can be either a dataset or a data dictionary.

Value

Either a tibble, identifying the dataset, or a list of tibble(s) identifying a data dictionary, depending which the input refers to.

Details

A data dictionary-like structure must be a list of at least one or two data-frame or data-frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated dataset. The 'Variables' component must contain at least 'name' column and the 'Categories' component must at least contain 'variable' and 'name' columns to be usable in any function of the package. To be considered as a minimum (workable) data dictionary, it must also have unique and non-null entries in 'name' column and the combination 'name'/'variable' must also be unique in 'Categories'. In addition, the data dictionary may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', 'missing' and 'label(:xx)', and/or any taxonomy provided.

A dataset must be a data-frame or data-frame extension (e.g. a tibble) and can be associated to a data dictionary. If not, a minimum workable data dictionary can always be generated, when any column will be reported, and any factor column will be analysed as categorical variable (the column 'levels' will be created for that. In addition, the dataset may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, or id columns declaration (which full completeness is mandatory.

See also

Examples

{

# use DEMO_files provided by the package

data_dict <- DEMO_files$dd_TOKYO_format_maelstrom_tagged
data_dict$Variables$valueType <- "text"
dataset <- DEMO_files$dataset_TOKYO

###### Example 1: The valueType of a dataset can be adjusted. each column is
# evaluated as whole, and the best valueType match found is applied. If 
# there is no better match found, the column is left as it is.
valueType_self_adjust(dataset)

#' ###### Example 2: The valueType present in a data dictionary can be 
# adjusted (only for categorical variables). Each categorical variable is
# evaluated as whole (with the values present in the 'name' column in the 
# 'Categories' component), and the best valueType match found is applied. If 
# there is no better match found, the valueType is left as it is.
valueType_self_adjust(data_dict)

}
#> $Variables
#> # A tibble: 9 × 15
#>   index name  `label:en` `description:en` valueType unit  Mlstr_additional::So…¹
#>   <dbl> <chr> <chr>      <chr>            <chr>     <chr> <chr>                 
#> 1     1 part… id of the… id of the parti… text      NA    Questionnaire         
#> 2     2 gndr  gndr       gender of the p… text      NA    Questionnaire         
#> 3     3 heig… height     height of the p… text      cm    Physical_measure      
#> 4     4 weig… weight_ms  weight of the p… integer   kg    Physical_measure      
#> 5     5 weig… weight_dc  weight of the p… text      kg    Physical_measure      
#> 6     6 dob   dob        date of birth o… text      years Questionnaire         
#> 7     7 prg_… prg_ever   whether the par… integer   NA    Questionnaire         
#> 8     8 empty empty      empty column     text      NA    Questionnaire         
#> 9     9 open… opentext   open text        text      NA    Questionnaire         
#> # ℹ abbreviated name: ¹​`Mlstr_additional::Source`
#> # ℹ 8 more variables: `Mlstr_additional::Target` <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <lgl>,
#> #   `Mlstr_area::2` <chr>, `Mlstr_area::2.term` <chr>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>
#> 
#> $Categories
#> # A tibble: 11 × 4
#>    variable  name   `label:en`            missing
#>    <chr>     <chr>  <chr>                 <lgl>  
#>  1 gndr      Male   Male                  FALSE  
#>  2 gndr      Female Female                FALSE  
#>  3 gndr      -77    Don’t want to answer  TRUE   
#>  4 weight_ms -88    Don’t want to answer  TRUE   
#>  5 weight_ms -99    Don’t know            TRUE   
#>  6 prg_ever  0      never pregnant        FALSE  
#>  7 prg_ever  1      pregnant once or more FALSE  
#>  8 prg_ever  2      currently pregnant    FALSE  
#>  9 prg_ever  8      Don’t want to answer  TRUE   
#> 10 prg_ever  9      Don’t know            TRUE   
#> 11 prg_ever  -7     not applicable        TRUE   
#>